Split the example application code into multiple files
Also fixes some mem leaks in long-living "group" variables
