+++ app/app_main.c
@@ -1,5 +1,5 @@
 /*****************************************************************************
-* Copyright (c) 2016, Cisco Systems, Inc.
+* Copyright (c) 2019, Cisco Systems, Inc.
 * All rights reserved.
 
 * Redistribution and use in source and binary forms, with or without modification,
@@ -31,54 +31,24 @@
  * It will default to 127.0.0.1 port 443 if no arguments are given.
  */
 #include <stdio.h>
-#include <string.h>
-#include <stdlib.h>
-#ifdef WIN32
-#include <io.h>
-#else
-#include <unistd.h>
-#endif
-#include <fcntl.h>
-#include "acvp/acvp.h"
-#include "safe_lib.h"
-
-#include <openssl/ossl_typ.h>
-#include <openssl/evp.h>
-#include <openssl/aes.h>
-#include <openssl/obj_mac.h>
-#include <openssl/err.h>
-#include <openssl/hmac.h>
-#include <openssl/ec.h>
-#include <openssl/ecdh.h>
-#include <openssl/cmac.h>
 #include <openssl/rsa.h>
 #include <openssl/bn.h>
-#include <openssl/rand.h>
-#include <openssl/ecdsa.h>
-#include <openssl/ec.h>
 
-#ifdef OPENSSL_KDF_SUPPORT
-#include <openssl/kdf.h>
-#endif
+#include "app_lcl.h"
+#include "safe_lib.h"
 
 #ifdef ACVP_NO_RUNTIME
-#include <openssl/dsa.h>
-#include "app_lcl.h"
-#include <openssl/fips_rand.h>
-#include <openssl/fips.h>
+# include "app_fips_lcl.h"
+# include "app_fips_init_lcl.h"
 extern int fips_selftest_fail;
 extern int fips_mode;
 #endif
-static ACVP_RESULT totp(char **token, int token_max);
+
 static int enable_aes(ACVP_CTX *ctx);
 static int enable_tdes(ACVP_CTX *ctx);
 static int enable_hash(ACVP_CTX *ctx);
 static int enable_cmac(ACVP_CTX *ctx);
 static int enable_hmac(ACVP_CTX *ctx);
-
-#ifdef OPENSSL_KDF_SUPPORT
-static int enable_kdf(ACVP_CTX *ctx);
-#endif
 #ifdef ACVP_NO_RUNTIME
 static int enable_dsa(ACVP_CTX *ctx);
 static int enable_rsa(ACVP_CTX *ctx);
@@ -87,71 +57,9 @@ static int enable_drbg(ACVP_CTX *ctx);
 static int enable_kas_ecc(ACVP_CTX *ctx);
 static int enable_kas_ffc(ACVP_CTX *ctx);
 #endif
-
-static int app_aes_handler_aead(ACVP_TEST_CASE *test_case);
-static int app_aes_handler(ACVP_TEST_CASE *test_case);
-static int app_des_handler(ACVP_TEST_CASE *test_case);
-static int app_sha_handler(ACVP_TEST_CASE *test_case);
-static int app_hmac_handler(ACVP_TEST_CASE *test_case);
-static int app_cmac_handler(ACVP_TEST_CASE *test_case);
-static int app_aes_keywrap_handler(ACVP_TEST_CASE *test_case);
-
 #ifdef OPENSSL_KDF_SUPPORT
-static int app_kdf135_tls_handler(ACVP_TEST_CASE *test_case);
-static int app_kdf135_snmp_handler(ACVP_TEST_CASE *test_case);
-static int app_kdf135_ssh_handler(ACVP_TEST_CASE *test_case);
-static int app_kdf135_srtp_handler(ACVP_TEST_CASE *test_case);
-static int app_kdf135_ikev2_handler(ACVP_TEST_CASE *test_case);
-static int app_kdf135_ikev1_handler(ACVP_TEST_CASE *test_case);
-static int app_kdf135_x963_handler(ACVP_TEST_CASE *test_case);
-static int app_kdf108_handler(ACVP_TEST_CASE *test_case);
+static int enable_kdf(ACVP_CTX *ctx);
 #endif
-#ifdef ACVP_NO_RUNTIME
-static int app_dsa_handler(ACVP_TEST_CASE *test_case);
-static int app_kas_ecc_handler(ACVP_TEST_CASE *test_case);
-static int app_kas_ffc_handler(ACVP_TEST_CASE *test_case);
-static int app_drbg_handler(ACVP_TEST_CASE *test_case);
-static int app_rsa_keygen_handler(ACVP_TEST_CASE *test_case);
-static int app_rsa_sig_handler(ACVP_TEST_CASE *test_case);
-static int app_ecdsa_handler(ACVP_TEST_CASE *test_case);
-#endif
-
-#define JSON_FILENAME_LENGTH 24
-#define DEFAULT_SERVER "127.0.0.1"
-#define DEFAULT_PORT 443
-#define DEFAULT_CA_CHAIN "certs/acvp-private-root-ca.crt.pem"
-#define DEFAULT_CERT "certs/my-client-cert.pem"
-#define DEFAULT_KEY "certs/my-client-key.pem"
-
-#define TLS_MD_MASTER_SECRET_CONST              "master secret"
-#define TLS_MD_MASTER_SECRET_CONST_SIZE         13
-#define TLS_MD_KEY_EXPANSION_CONST              "key expansion"
-#define TLS_MD_KEY_EXPANSION_CONST_SIZE         13
-
-typedef struct app_config {
-    ACVP_LOG_LVL level;
-    int sample;
-    int dev;
-    int json;
-    char json_file[JSON_FILENAME_LENGTH + 1];
-
-    /*
-     * Algorithm Flags
-     * 0 is off, 1 is on
-     */
-    int aes; int tdes;
-    int hash; int cmac;
-    int hmac;
-    /* These require the fom */
-#ifdef ACVP_NO_RUNTIME
-    int dsa; int rsa;
-    int drbg; int ecdsa;
-    int kas_ecc; int kas_ffc;
-#endif
-#ifdef OPENSSL_KDF_SUPPORT
-    int kdf;
-#endif
-} APP_CONFIG;
 
 char *server;
 int port;
@@ -162,34 +70,11 @@ char *path_segment;
 char *api_context;
 char value[] = "same";
 
-static EVP_CIPHER_CTX *glb_cipher_ctx = NULL; /* need to maintain across calls for MCT */
-
-/* RSA group values */
-int rsa_current_tg = 0;
-BIGNUM *group_n = NULL;
-RSA *group_rsa = NULL;
-
-/* ECDSA group values */
-int ecdsa_current_tg = 0;
-BIGNUM *ecdsa_group_Qx = NULL;
-BIGNUM *ecdsa_group_Qy = NULL;
-EC_KEY *ecdsa_group_key = NULL;
-
-/* DSA group values */
-DSA *group_dsa = NULL;
-int dsa_current_keygen_tg = 0;
-int dsa_current_siggen_tg = 0;
-BIGNUM *group_p = NULL;
-BIGNUM *group_q = NULL;
-BIGNUM *group_g = NULL;
-BIGNUM *group_pub_key = NULL;
-
 #define CHECK_ENABLE_CAP_RV(rv) \
     if (rv != ACVP_SUCCESS) { \
         printf("Failed to register capability with libacvp (rv=%d: %s)\n", rv, acvp_lookup_error_string(rv)); \
         goto end; \
     }
-
 
 /*
  * Read the operational parameters from the various environment
@@ -234,561 +119,23 @@ static void setup_session_parameters() {
  * libacvp calls this function to for debugs, warnings,
  * and errors.
  */
-ACVP_RESULT progress(char *msg) {
+static ACVP_RESULT progress(char *msg) {
     printf("%s", msg);
     return ACVP_SUCCESS;
 }
 
-#define ANSI_COLOR_RED "\x1b[31m"
-#define ANSI_COLOR_YELLOW "\x1b[33m"
-#define ANSI_COLOR_RESET "\x1b[0m"
+static void app_cleanup(ACVP_CTX *ctx) {
+    // Routines for libacvp
+    acvp_cleanup(ctx);
 
-static void print_usage(int err) {
-    if (err) {
-        printf("\nInvalid usage...\n");
-    } else {
-        printf("\n===========================");
-        printf("\n===== ACVP_APP USAGE ======");
-        printf("\n===========================\n");
-    }
-    printf("This program does not require any argument, however logging level can be\n");
-    printf("controlled using:\n");
-    printf("      --none\n");
-    printf("      --error\n");
-    printf("      --warn\n");
-    printf("      --status(default)\n");
-    printf("      --info\n");
-    printf("      --verbose\n");
-    printf("      --version\n");
-    printf("\n");
-    printf("Algorithm Test Suites:\n");
-    printf("      --aes(default)\n");
-    printf("      --no_aes\n");
-    printf("      --tdes(default)\n");
-    printf("      --no_tdes\n");
-    printf("      --hash(default)\n");
-    printf("      --no_hash\n");
-    printf("      --cmac(default)\n");
-    printf("      --no_cmac\n");
-    printf("      --hmac(default)\n");
-    printf("      --no_hmac\n");
-#ifdef OPENSSL_KDF_SUPPORT
-    printf("      --kdf\n");
-    printf("      --no_kdf(default)\n");
-#endif
+    // Routines for this application
+    app_aes_cleanup();
+    app_des_cleanup();
 #ifdef ACVP_NO_RUNTIME
-    printf("      --dsa\n");
-    printf("      --no_dsa(default)\n");
-    printf("      --rsa\n");
-    printf("      --no_rsa(default)\n");
-    printf("      --ecdsa\n");
-    printf("      --no_ecdsa(default)\n");
-    printf("      --drbg\n");
-    printf("      --no_drbg(default)\n");
-    printf("      --kas_ecc\n");
-    printf("      --no_kas_ecc(default)\n");
-    printf("      --kas_ffc\n");
-    printf("      --no_kas_ffc(default)\n");
+    app_dsa_cleanup();
+    app_rsa_cleanup();
+    app_ecdsa_cleanup();
 #endif
-    printf("\n");
-    printf("To register a formatted JSON file use:\n");
-    printf("      --json <file>\n");
-    printf("\n");
-    printf("If you are running a sample registration (querying for correct answers\n");
-    printf("in addition to the normal registration flow) use:\n");
-    printf("      --sample\n");
-    printf("\n");
-    printf("If you want to include \"debugRequest\" in your registration, use:\n");
-    printf("      --dev\n");
-    printf("\n");
-    printf("In addition some options are passed to acvp_app using\n");
-    printf("environment variables.  The following variables can be set:\n\n");
-    printf("    ACV_SERVER (when not set, defaults to %s)\n", DEFAULT_SERVER);
-    printf("    ACV_PORT (when not set, defaults to %d)\n", DEFAULT_PORT);
-    printf("    ACV_URI_PREFIX (when not set, defaults to null)\n");
-    printf("    ACV_CA_FILE (when not set, defaults to %s)\n", DEFAULT_CA_CHAIN);
-    printf("    ACV_CERT_FILE (when not set, defaults to %s)\n", DEFAULT_CERT);
-    printf("    ACV_KEY_FILE (when not set, defaults to %s)\n", DEFAULT_KEY);
-    printf("    ACV_TOTP_SEED (when not set, client will not use Two-factor authentication)\n\n");
-    printf("The CA certificates, cert and key should be PEM encoded. There should be no\n");
-    printf("password on the key file.\n");
-}
-
-static int cli_alg_option(int *alg,
-                          int *op_status,
-                          int enable,
-                          char *enable_str,
-                          char *disable_str) {
-#define OP_DISABLE 1
-#define OP_ENABLE 2
-    if (enable) {
-        /*
-         * Trying to enable this algorithm.
-         * Check to see if algorithm has been disabled already.
-         */
-        if (*op_status == OP_DISABLE) {
-            printf(ANSI_COLOR_RED "Command error... [%s]"ANSI_COLOR_RESET
-                   "\nAlgorithm already disabled by \"%s\"."
-                   "\nPlease give only 1 of these options at a time.\n",
-                   enable_str, disable_str);
-            return 1;
-        }
-        *op_status = OP_ENABLE;
-        *alg = 1;
-    } else {
-        /*
-         * Trying to disable this algorithm.
-         * Check to see if algorithm has been disabled already.
-         */
-        if (*op_status == OP_ENABLE) {
-            printf(ANSI_COLOR_RED "Command error... [%s]"ANSI_COLOR_RESET
-                   "\nAlgorithm already enabled by \"%s\"."
-                   "\nPlease give only 1 of these options at a time.\n",
-                   disable_str, enable_str);
-            return 1;
-        }
-        *op_status = OP_DISABLE;
-        *alg = 0;
-    }
-
-    return 0;
-}
-
-static void default_config(APP_CONFIG *cfg) {
-    cfg->level = ACVP_LOG_LVL_STATUS;
-    cfg->aes = 1;
-    cfg->tdes = 1;
-    cfg->hash = 1;
-    cfg->cmac = 1;
-    cfg->hmac = 1;
-}
-
-static int ingest_cli(APP_CONFIG *cfg, int argc, char **argv) {
-    char *log_lvl = NULL;
-    int diff = 0;
-    int aes_status = 0, tdes_status = 0,
-        hash_status = 0, cmac_status = 0,
-        hmac_status = 0;
-
-#ifdef ACVP_NO_RUNTIME
-    int dsa_status = 0, rsa_status = 0,
-        drbg_status = 0, ecdsa_status = 0,
-        kas_ecc_status = 0, kas_ffc_status = 0;
-#endif
-#ifdef OPENSSL_KDF_SUPPORT
-    int kdf_status = 0;
-#endif
-
-#define ALG_DISABLE 0
-#define ALG_ENABLE 1
-#define OPTION_STR_MAX 16
-
-    /* Set the default configuration values */
-    default_config(cfg);
-
-    argv++;
-    argc--;
-    while (argc >= 1) {
-        /* version option used by itself, ignore remaining command line */
-        strcmp_s("--version", strnlen_s("--version", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            printf("\nACVP library version(protocol version): %s(%s)\n", acvp_version(), acvp_protocol_version());
-            return 1;
-        }
-
-        strcmp_s("--help", strnlen_s("--help", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            print_usage(0);
-            return 1;
-        }
-
-        strcmp_s("--info", strnlen_s("--info", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            if (log_lvl) {
-                printf(ANSI_COLOR_RED "Command error... [%s]"ANSI_COLOR_RESET
-                       "\nLog Level already set to \"%s\"."
-                       "\nOnly 1 Log Level can be specified.\n", "--info", log_lvl);
-                print_usage(1);
-                return 1;
-            }
-            cfg->level = ACVP_LOG_LVL_INFO;
-            log_lvl = "info";
-            goto next;
-        }
-
-        strcmp_s("--status", strnlen_s("--status", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            if (log_lvl) {
-                printf(ANSI_COLOR_RED "Command error... [%s]"ANSI_COLOR_RESET
-                       "\nLog Level already set to \"%s\"."
-                       "\nOnly 1 Log Level can be specified.\n", "--status", log_lvl);
-                print_usage(1);
-                return 1;
-            }
-            cfg->level = ACVP_LOG_LVL_STATUS;
-            log_lvl = "status";
-            goto next;
-        }
-
-        strcmp_s("--warn", strnlen_s("--warn", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            if (log_lvl) {
-                printf(ANSI_COLOR_RED "Command error... [%s]"ANSI_COLOR_RESET
-                       "\nLog Level already set to \"%s\"."
-                       "\nOnly 1 Log Level can be specified.\n", "--warn", log_lvl);
-                print_usage(1);
-                return 1;
-            }
-            cfg->level = ACVP_LOG_LVL_WARN;
-            log_lvl = "warn";
-            goto next;
-        }
-
-        strcmp_s("--error", strnlen_s("--error", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            if (log_lvl) {
-                printf(ANSI_COLOR_RED "Command error... [%s]"ANSI_COLOR_RESET
-                       "\nLog Level already set to \"%s\"."
-                       "\nOnly 1 Log Level can be specified.\n", "--error", log_lvl);
-                print_usage(1);
-                return 1;
-            }
-            cfg->level = ACVP_LOG_LVL_ERR;
-            log_lvl = "error";
-            goto next;
-        }
-
-        strcmp_s("--none", strnlen_s("--none", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            if (log_lvl) {
-                printf(ANSI_COLOR_RED "Command error... [%s]"ANSI_COLOR_RESET
-                       "\nLog Level already set to \"%s\"."
-                       "\nOnly 1 Log Level can be specified.\n", "--none", log_lvl);
-                print_usage(1);
-                return 1;
-            }
-            cfg->level = ACVP_LOG_LVL_NONE;
-            log_lvl = "none";
-            goto next;
-        }
-
-        strcmp_s("--verbose", strnlen_s("--verbose", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            if (log_lvl) {
-                printf(ANSI_COLOR_RED "Command error... [%s]"ANSI_COLOR_RESET
-                       "\nLog Level already set to \"%s\"."
-                       "\nOnly 1 Log Level can be specified.\n", "--verbose", log_lvl);
-                print_usage(1);
-                return 1;
-            }
-            cfg->level = ACVP_LOG_LVL_VERBOSE;
-            log_lvl = "verbose";
-            goto next;
-        }
-
-        strcmp_s("--json", strnlen_s("--json", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            int filename_len = 0;
-
-            cfg->json = 1;
-            argc--;
-            argv++;
-
-            if (*argv == NULL) {
-                printf(ANSI_COLOR_RED "Command error... [%s]"ANSI_COLOR_RESET
-                       "\nMissing <file>.\n", "--json");
-                print_usage(1);
-                return 1;
-            }
-
-            filename_len = strnlen_s(*argv, JSON_FILENAME_LENGTH + 1);
-            if (filename_len > JSON_FILENAME_LENGTH) {
-                printf(ANSI_COLOR_RED "Command error... [%s]"ANSI_COLOR_RESET
-                       "\nThe <file> \"%s\", has a name that is too long."
-                       "\nMax allowed <file> name length is (%d).\n",
-                       "--json", *argv, JSON_FILENAME_LENGTH);
-                print_usage(1);
-                return 1;
-            }
-
-            strcpy_s(cfg->json_file, JSON_FILENAME_LENGTH + 1, *argv);
-        }
-
-        strcmp_s("--sample", strnlen_s("--sample", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            cfg->sample = 1;
-            goto next;
-        }
-
-        strcmp_s("--dev", strnlen_s("--dev", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            cfg->dev = 1;
-            goto next;
-        }
-
-        strcmp_s("--aes", strnlen_s("--aes", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            if (cli_alg_option(&cfg->aes, &aes_status, ALG_ENABLE,
-                               "--aes", "--no_aes")) return 1;
-            goto next;
-        }
-
-        strcmp_s("--no_aes", strnlen_s("--no_aes", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            if (cli_alg_option(&cfg->aes, &aes_status, ALG_DISABLE,
-                               "--aes", "--no_aes")) return 1;
-            goto next;
-        }
-
-        strcmp_s("--tdes", strnlen_s("--tdes", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            if (cli_alg_option(&cfg->tdes, &tdes_status, ALG_ENABLE,
-                               "--tdes", "--no_tdes")) return 1;
-            goto next;
-        }
-
-        strcmp_s("--no_tdes", strnlen_s("--no_tdes", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            if (cli_alg_option(&cfg->tdes, &tdes_status, ALG_DISABLE,
-                               "--tdes", "--no_tdes")) return 1;
-            goto next;
-        }
-
-        strcmp_s("--hash", strnlen_s("--hash", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            if (cli_alg_option(&cfg->hash, &hash_status, ALG_ENABLE,
-                               "--hash", "--no_hash")) return 1;
-            goto next;
-        }
-
-        strcmp_s("--no_hash", strnlen_s("--no_hash", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            if (cli_alg_option(&cfg->hash, &hash_status, ALG_DISABLE,
-                               "--hash", "--no_hash")) return 1;
-            goto next;
-        }
-
-        strcmp_s("--cmac", strnlen_s("--cmac", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            if (cli_alg_option(&cfg->cmac, &cmac_status, ALG_ENABLE,
-                               "--cmac", "--no_cmac")) return 1;
-            goto next;
-        }
-
-        strcmp_s("--no_cmac", strnlen_s("--no_cmac", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            if (cli_alg_option(&cfg->cmac, &cmac_status, ALG_DISABLE,
-                               "--cmac", "--no_cmac")) return 1;
-            goto next;
-        }
-
-        strcmp_s("--hmac", strnlen_s("--hmac", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            if (cli_alg_option(&cfg->hmac, &hmac_status, ALG_ENABLE,
-                               "--hmac", "--no_hmac")) return 1;
-            goto next;
-        }
-
-        strcmp_s("--no_hmac", strnlen_s("--no_hmac", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-            if (cli_alg_option(&cfg->hmac, &hmac_status, ALG_DISABLE,
-                               "--hmac", "--no_hmac")) return 1;
-            goto next;
-        }
-
-        strcmp_s("--kdf", strnlen_s("--kdf", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-#ifdef OPENSSL_KDF_SUPPORT
-            if (cli_alg_option(&cfg->kdf, &kdf_status, ALG_ENABLE,
-                               "--kdf", "--no_kdf")) return 1;
-#else
-            printf(ANSI_COLOR_RED "Command error... [%s]"ANSI_COLOR_RESET
-                   "\nMissing compile flag -DOPENSSL_KDF_SUPPORT"
-                   "\nThis option will have no effect.\n", "--kdf");
-#endif
-            goto next;
-        }
-
-        strcmp_s("--no_kdf", strnlen_s("--no_kdf", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-#ifdef OPENSSL_KDF_SUPPORT
-            if (cli_alg_option(&cfg->kdf, &kdf_status, ALG_DISABLE,
-                               "--kdf", "--no_kdf")) return 1;
-#else
-            printf(ANSI_COLOR_RED "Command error... [%s]"ANSI_COLOR_RESET
-                   "\nMissing compile flag -DOPENSSL_KDF_SUPPORT"
-                   "\nThis option will have no effect.\n", "--no_kdf");
-#endif
-            goto next;
-        }
-
-        strcmp_s("--dsa", strnlen_s("--dsa", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-#ifdef ACVP_NO_RUNTIME
-            if (cli_alg_option(&cfg->dsa, &dsa_status, ALG_ENABLE,
-                               "--dsa", "--no_dsa")) return 1;
-#else
-            printf(ANSI_COLOR_RED "Command error... [%s]"ANSI_COLOR_RESET
-                   "\nMissing compile flag -DACVP_NO_RUNTIME"
-                   "\nThis option will have no effect.\n", "--dsa");
-#endif
-            goto next;
-        }
-
-        strcmp_s("--no_dsa", strnlen_s("--no_dsa", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-#ifdef ACVP_NO_RUNTIME
-            if (cli_alg_option(&cfg->dsa, &dsa_status, ALG_DISABLE,
-                               "--dsa", "--no_dsa")) return 1;
-#else
-            printf(ANSI_COLOR_RED "Command error... [%s]"ANSI_COLOR_RESET
-                   "\nMissing compile flag -DACVP_NO_RUNTIME"
-                   "\nThis option will have no effect.\n", "--no_dsa");
-#endif
-            goto next;
-        }
-
-        strcmp_s("--rsa", strnlen_s("--rsa", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-#ifdef ACVP_NO_RUNTIME
-            if (cli_alg_option(&cfg->rsa, &rsa_status, ALG_ENABLE,
-                               "--rsa", "--no_rsa")) return 1;
-#else
-            printf(ANSI_COLOR_RED "Command error... [%s]"ANSI_COLOR_RESET
-                   "\nMissing compile flag -DACVP_NO_RUNTIME"
-                   "\nThis option will have no effect.\n", "--rsa");
-#endif
-            goto next;
-        }
-
-        strcmp_s("--no_rsa", strnlen_s("--no_rsa", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-#ifdef ACVP_NO_RUNTIME
-            if (cli_alg_option(&cfg->rsa, &rsa_status, ALG_DISABLE,
-                               "--rsa", "--no_rsa")) return 1;
-#else
-            printf(ANSI_COLOR_RED "Command error... [%s]"ANSI_COLOR_RESET
-                   "\nMissing compile flag -DACVP_NO_RUNTIME"
-                   "\nThis option will have no effect.\n", "--no_rsa");
-#endif
-            goto next;
-        }
-
-        strcmp_s("--drbg", strnlen_s("--drbg", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-#ifdef ACVP_NO_RUNTIME
-            if (cli_alg_option(&cfg->drbg, &drbg_status, ALG_ENABLE,
-                               "--drbg", "--no_drbg")) return 1;
-#else
-            printf(ANSI_COLOR_RED "Command error... [%s]"ANSI_COLOR_RESET
-                   "\nMissing compile flag -DACVP_NO_RUNTIME"
-                   "\nThis option will have no effect\n", "--drbg");
-#endif
-            goto next;
-        }
-
-        strcmp_s("--no_drbg", strnlen_s("--no_drbg", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-#ifdef ACVP_NO_RUNTIME
-            if (cli_alg_option(&cfg->drbg, &drbg_status, ALG_DISABLE,
-                               "--drbg", "--no_drbg")) return 1;
-#else
-            printf(ANSI_COLOR_RED "Command error... [%s]"ANSI_COLOR_RESET
-                   "\nMissing compile flag -DACVP_NO_RUNTIME"
-                   "\nTHis option will have no effect.\n", "--no_drbg");
-#endif
-            goto next;
-        }
-
-        strcmp_s("--ecdsa", strnlen_s("--ecdsa", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-#ifdef ACVP_NO_RUNTIME
-            if (cli_alg_option(&cfg->ecdsa, &ecdsa_status, ALG_ENABLE,
-                               "--ecdsa", "--no_ecdsa")) return 1;
-#else
-            printf(ANSI_COLOR_YELLOW "Command warning... [%s]"ANSI_COLOR_RESET
-                   "\nMissing compile flag -DACVP_NO_RUNTIME"
-                   "\nThis option will have no effect.\n", "--ecdsa");
-#endif
-            goto next;
-        }
-
-        strcmp_s("--no_ecdsa", strnlen_s("--no_ecdsa", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-#ifdef ACVP_NO_RUNTIME
-            if (cli_alg_option(&cfg->ecdsa, &ecdsa_status, ALG_DISABLE,
-                               "--ecdsa", "--no_ecdsa")) return 1;
-#else
-            printf(ANSI_COLOR_YELLOW "Command warning... [%s]"ANSI_COLOR_RESET
-                   "\nMissing compile flag -DACVP_NO_RUNTIME"
-                   "\nThis options will have no effect.\n", "--no_ecdsa");
-#endif
-            goto next;
-        }
-
-        strcmp_s("--kas_ecc", strnlen_s("--kas_ecc", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-#ifdef ACVP_NO_RUNTIME
-            if (cli_alg_option(&cfg->kas_ecc, &kas_ecc_status, ALG_ENABLE,
-                               "--kas_ecc", "--no_kas_ecc")) return 1;
-#else
-            printf(ANSI_COLOR_YELLOW "Command warning... [%s]"ANSI_COLOR_RESET
-                   "\nMissing compile flag -DACVP_NO_RUNTIME"
-                   "\nThis option will have no effect.\n", "--kas_ecc");
-#endif
-            goto next;
-        }
-
-        strcmp_s("--no_kas_ecc", strnlen_s("--no_kas_ecc", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-#ifdef ACVP_NO_RUNTIME
-            if (cli_alg_option(&cfg->kas_ecc, &kas_ecc_status, ALG_DISABLE,
-                               "--kas_ecc", "--no_kas_ecc")) return 1;
-#else
-            printf(ANSI_COLOR_YELLOW "Command warning... [%s]"ANSI_COLOR_RESET
-                   "\nMissing compile flag -DACVP_NO_RUNTIME"
-                   "\nThis option will have no effect.\n", "--no_kas_ecc");
-#endif
-            goto next;
-        }
-
-        strcmp_s("--kas_ffc", strnlen_s("--kas_ffc", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-#ifdef ACVP_NO_RUNTIME
-            if (cli_alg_option(&cfg->kas_ffc, &kas_ffc_status, ALG_ENABLE,
-                               "--kas_ffc", "--no_kas_ffc")) return 1;
-#else
-            printf(ANSI_COLOR_YELLOW "Command warning... [%s]"ANSI_COLOR_RESET
-                   "\nMissing compile flag -DACVP_NO_RUNTIME"
-                   "\nThis option will have no effect.\n", "--kas_ffc");
-#endif
-            goto next;
-        }
-
-        strcmp_s("--no_kas_ffc", strnlen_s("--no_kas_ffc", OPTION_STR_MAX), *argv, &diff);
-        if (!diff) {
-#ifdef ACVP_NO_RUNTIME
-            if (cli_alg_option(&cfg->kas_ffc, &kas_ffc_status, ALG_DISABLE,
-                               "--kas_ffc", "--no_kas_ffc")) return 1;
-#else
-            printf(ANSI_COLOR_YELLOW "Command warning... [%s]"ANSI_COLOR_RESET
-                   "\nMissing compile flag -DACVP_NO_RUNTIME"
-                   "\nThis option will have no effect.\n", "--no_kas_ffc");
-#endif
-            goto next;
-        }
-
-        /* If you get here, the command wasn't recognized */
-        printf("Command error... Option not recognized: \"%s\"", *argv);
-        print_usage(1);
-        return 1;
-next:
-        argv++;
-        argc--;
-    }
-    printf("\n");
-
-    return 0;
 }
 
 int main(int argc, char **argv) {
@@ -809,13 +156,6 @@ int main(int argc, char **argv) {
     fips_algtest_init_nofips();
 #endif
 
-    if (glb_cipher_ctx == NULL) {
-        glb_cipher_ctx = EVP_CIPHER_CTX_new();
-        if (glb_cipher_ctx == NULL) {
-            printf("Failed to allocate global cipher_ctx");
-            goto end;
-        }
-    }
     setup_session_parameters();
 
     /*
@@ -1030,22 +370,11 @@ int main(int argc, char **argv) {
     }
 
 end:
-    if (glb_cipher_ctx) EVP_CIPHER_CTX_free(glb_cipher_ctx);
-    /* free RSA group vals */
-    if (group_rsa) RSA_free(group_rsa);
-    if (group_n) BN_free(group_n);
-    /* free DSA group vals */
-    if (group_dsa) DSA_free(group_dsa);
-    if (group_p) BN_free(group_p);
-    if (group_q) BN_free(group_q);
-    if (group_g) BN_free(group_g);
-    if (group_pub_key) BN_free(group_pub_key);
-    /* free ECDSA group vals */
-    if (ecdsa_group_Qx) BN_free(ecdsa_group_Qx);
-    if (ecdsa_group_Qy) BN_free(ecdsa_group_Qy);
-    if (ecdsa_group_key) EC_KEY_free(ecdsa_group_key);
-    /* Free all memory associated with libacvp */
-    rv = acvp_cleanup(ctx);
+    /*
+     * Free all memory associated with
+     * both the application and libacvp.
+     */
+    app_cleanup(ctx);
 
     return rv;
 }
@@ -1702,8 +1031,6 @@ end:
 }
 
 #ifdef OPENSSL_KDF_SUPPORT
-#define ENGID1 "800002B805123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456"
-#define ENGID2 "000002b87766554433221100"
 static int enable_kdf(ACVP_CTX *ctx) {
     ACVP_RESULT rv = ACVP_SUCCESS;
     int i, flags = 0;
@@ -1797,6 +1124,7 @@ static int enable_kdf(ACVP_CTX *ctx) {
     rv = acvp_cap_kdf135_ikev2_set_parm(ctx, ACVP_KDF_HASH_ALG, ACVP_SHA1);
     CHECK_ENABLE_CAP_RV(rv);
 
+#if 0 // Disabled for now
     rv = acvp_cap_kdf135_ikev1_enable(ctx, &app_kdf135_ikev1_handler);
     CHECK_ENABLE_CAP_RV(rv);
     rv = acvp_cap_set_prereq(ctx, ACVP_KDF135_IKEV1, ACVP_PREREQ_SHA, value);
@@ -1840,6 +1168,7 @@ static int enable_kdf(ACVP_CTX *ctx) {
     CHECK_ENABLE_CAP_RV(rv);
     rv = acvp_cap_kdf135_x963_set_parm(ctx, ACVP_KDF_X963_SHARED_INFO_LEN, 1024);
     CHECK_ENABLE_CAP_RV(rv);
+#endif
 
     /*
      * KDF108 Counter Mode
@@ -2163,13 +1492,13 @@ static int enable_rsa(ACVP_CTX *ctx) {
     BIGNUM *expo = NULL;
     char *expo_str = NULL;
 
-    expo = FIPS_bn_new();
+    expo = BN_new();
     if (!expo || !BN_set_word(expo, RSA_F4)) {
         printf("oh no\n");
         return 1;
     }
     expo_str = BN_bn2hex(expo);
-    FIPS_bn_free(expo);
+    BN_free(expo);
 
     /*
      * Enable RSA keygen...
@@ -2833,2800 +2162,3 @@ end:
 }
 #endif
 
-static int app_des_handler(ACVP_TEST_CASE *test_case) {
-    ACVP_SYM_CIPHER_TC      *tc;
-    EVP_CIPHER_CTX *cipher_ctx;
-    const EVP_CIPHER        *cipher;
-    int ct_len, pt_len;
-    unsigned char *iv = 0;
-
-    if (!test_case) {
-        return 1;
-    }
-
-    tc = test_case->tc.symmetric;
-
-    /*
-     * We only support 3 key DES
-     */
-    if (tc->key_len != 192) {
-        printf("Unsupported DES key length\n");
-        return 1;
-    }
-
-    /* Begin encrypt code section */
-    cipher_ctx = glb_cipher_ctx;
-
-    switch (tc->cipher) {
-    case ACVP_TDES_ECB:
-        cipher = EVP_des_ede3_ecb();
-        break;
-    case ACVP_TDES_CBC:
-        iv = tc->iv;
-        cipher = EVP_des_ede3_cbc();
-        break;
-    case ACVP_TDES_OFB:
-        iv = tc->iv;
-        cipher = EVP_des_ede3_ofb();
-        break;
-    case ACVP_TDES_CFB64:
-        iv = tc->iv;
-        cipher = EVP_des_ede3_cfb64();
-        break;
-    case ACVP_TDES_CFB8:
-        iv = tc->iv;
-        cipher = EVP_des_ede3_cfb8();
-        break;
-    case ACVP_TDES_CFB1:
-        iv = tc->iv;
-        cipher = EVP_des_ede3_cfb1();
-        break;
-    case ACVP_TDES_CTR:
-    /*
-     * IMPORTANT: if this mode is supported in your crypto module,
-     * you will need to fill that out here. It is set to fall
-     * through as an unsupported mode.
-     */
-    default:
-        printf("Error: Unsupported DES mode requested by ACVP server\n");
-        return 1;
-
-        break;
-    }
-
-    /* If Monte Carlo we need to be able to init and then update
-     * one thousand times before we complete each iteration.
-     */
-    if (tc->test_type == ACVP_SYM_TEST_TYPE_MCT) {
-        const unsigned char *ctx_iv = NULL;
-
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-        ctx_iv = cipher_ctx->iv;
-#else
-        ctx_iv = EVP_CIPHER_CTX_iv(cipher_ctx);
-#endif
-
-#define SYM_IV_BYTE_MAX 128
-        if (tc->direction == ACVP_SYM_CIPH_DIR_ENCRYPT) {
-            if (tc->mct_index == 0) {
-                EVP_EncryptInit_ex(cipher_ctx, cipher, NULL, tc->key, iv);
-                EVP_CIPHER_CTX_set_padding(cipher_ctx, 0);
-            } else {
-                /* TDES needs the pre-operation IV returned */
-                memcpy_s(tc->iv_ret, SYM_IV_BYTE_MAX, ctx_iv, 8);
-            }
-            if (tc->cipher == ACVP_TDES_CFB1) {
-                EVP_CIPHER_CTX_set_flags(cipher_ctx, EVP_CIPH_FLAG_LENGTH_BITS);
-            }
-
-            EVP_EncryptUpdate(cipher_ctx, tc->ct, &ct_len, tc->pt, tc->pt_len);
-            tc->ct_len = ct_len;
-            /* TDES needs the post-operation IV returned */
-            memcpy_s(tc->iv_ret_after, SYM_IV_BYTE_MAX, ctx_iv, 8);
-        } else if (tc->direction == ACVP_SYM_CIPH_DIR_DECRYPT) {
-            if (tc->mct_index == 0) {
-                EVP_DecryptInit_ex(cipher_ctx, cipher, NULL, tc->key, iv);
-                EVP_CIPHER_CTX_set_padding(cipher_ctx, 0);
-            } else {
-                /* TDES needs the pre-operation IV returned */
-                memcpy_s(tc->iv_ret, SYM_IV_BYTE_MAX, ctx_iv, 8);
-            }
-            if (tc->cipher == ACVP_TDES_CFB1) {
-                EVP_CIPHER_CTX_set_flags(cipher_ctx, EVP_CIPH_FLAG_LENGTH_BITS);
-            }
-            EVP_DecryptUpdate(cipher_ctx, tc->pt, &pt_len, tc->ct, tc->ct_len);
-            tc->pt_len = pt_len;
-            /* TDES needs the post-operation IV returned */
-            memcpy_s(tc->iv_ret_after, SYM_IV_BYTE_MAX, ctx_iv, 8);
-        } else {
-            printf("Unsupported direction\n");
-            return 1;
-        }
-        if (tc->mct_index == 9999) {
-            EVP_CIPHER_CTX_cleanup(cipher_ctx);
-        }
-    } else {
-        if (tc->direction == ACVP_SYM_CIPH_DIR_ENCRYPT) {
-            EVP_EncryptInit_ex(cipher_ctx, cipher, NULL, tc->key, iv);
-            EVP_CIPHER_CTX_set_padding(cipher_ctx, 0);
-            if (tc->cipher == ACVP_TDES_CFB1) {
-                EVP_CIPHER_CTX_set_flags(cipher_ctx, EVP_CIPH_FLAG_LENGTH_BITS);
-            }
-            EVP_EncryptUpdate(cipher_ctx, tc->ct, &ct_len, tc->pt, tc->pt_len);
-            tc->ct_len = ct_len;
-            EVP_EncryptFinal_ex(cipher_ctx, tc->ct + ct_len, &ct_len);
-            tc->ct_len += ct_len;
-        } else if (tc->direction == ACVP_SYM_CIPH_DIR_DECRYPT) {
-            EVP_DecryptInit_ex(cipher_ctx, cipher, NULL, tc->key, iv);
-            EVP_CIPHER_CTX_set_padding(cipher_ctx, 0);
-            if (tc->cipher == ACVP_TDES_CFB1) {
-                EVP_CIPHER_CTX_set_flags(cipher_ctx, EVP_CIPH_FLAG_LENGTH_BITS);
-            }
-            EVP_DecryptUpdate(cipher_ctx, tc->pt, &pt_len, tc->ct, tc->ct_len);
-            tc->pt_len = pt_len;
-            EVP_DecryptFinal_ex(cipher_ctx, tc->pt + pt_len, &pt_len);
-            tc->pt_len += pt_len;
-        } else {
-            printf("Unsupported direction\n");
-            return 1;
-        }
-
-        EVP_CIPHER_CTX_cleanup(cipher_ctx);
-    }
-
-    return 0;
-}
-
-static int app_aes_handler(ACVP_TEST_CASE *test_case) {
-    ACVP_SYM_CIPHER_TC      *tc;
-    EVP_CIPHER_CTX *cipher_ctx;
-    const EVP_CIPHER        *cipher;
-    int ct_len, pt_len;
-    unsigned char *iv = 0;
-    /* assume fail at first */
-    int rv = 1;
-
-    if (!test_case) {
-        return rv;
-    }
-
-    tc = test_case->tc.symmetric;
-
-    /* Begin encrypt code section */
-    cipher_ctx = glb_cipher_ctx;
-    if ((tc->test_type != ACVP_SYM_TEST_TYPE_MCT)) {
-        EVP_CIPHER_CTX_init(cipher_ctx);
-    }
-
-    switch (tc->cipher) {
-    case ACVP_AES_ECB:
-        switch (tc->key_len) {
-        case 128:
-            cipher = EVP_aes_128_ecb();
-            break;
-        case 192:
-            cipher = EVP_aes_192_ecb();
-            break;
-        case 256:
-            cipher = EVP_aes_256_ecb();
-            break;
-        default:
-            printf("Unsupported AES key length\n");
-            return rv;
-
-            break;
-        }
-        break;
-    case ACVP_AES_CTR:
-        iv = tc->iv;
-        switch (tc->key_len) {
-        case 128:
-            cipher = EVP_aes_128_ctr();
-            break;
-        case 192:
-            cipher = EVP_aes_192_ctr();
-            break;
-        case 256:
-            cipher = EVP_aes_256_ctr();
-            break;
-        default:
-            printf("Unsupported AES key length\n");
-            return rv;
-
-            break;
-        }
-        break;
-    case ACVP_AES_CFB1:
-        iv = tc->iv;
-        switch (tc->key_len) {
-        case 128:
-            cipher = EVP_aes_128_cfb1();
-            break;
-        case 192:
-            cipher = EVP_aes_192_cfb1();
-            break;
-        case 256:
-            cipher = EVP_aes_256_cfb1();
-            break;
-        default:
-            printf("Unsupported AES key length\n");
-            return rv;
-
-            break;
-        }
-        break;
-    case ACVP_AES_CFB8:
-        iv = tc->iv;
-        switch (tc->key_len) {
-        case 128:
-            cipher = EVP_aes_128_cfb8();
-            break;
-        case 192:
-            cipher = EVP_aes_192_cfb8();
-            break;
-        case 256:
-            cipher = EVP_aes_256_cfb8();
-            break;
-        default:
-            printf("Unsupported AES key length\n");
-            return rv;
-
-            break;
-        }
-        break;
-    case ACVP_AES_CFB128:
-        iv = tc->iv;
-        switch (tc->key_len) {
-        case 128:
-            cipher = EVP_aes_128_cfb128();
-            break;
-        case 192:
-            cipher = EVP_aes_192_cfb128();
-            break;
-        case 256:
-            cipher = EVP_aes_256_cfb128();
-            break;
-        default:
-            printf("Unsupported AES key length\n");
-            return rv;
-
-            break;
-        }
-        break;
-    case ACVP_AES_OFB:
-        iv = tc->iv;
-        switch (tc->key_len) {
-        case 128:
-            cipher = EVP_aes_128_ofb();
-            break;
-        case 192:
-            cipher = EVP_aes_192_ofb();
-            break;
-        case 256:
-            cipher = EVP_aes_256_ofb();
-            break;
-        default:
-            printf("Unsupported AES key length\n");
-            return rv;
-
-            break;
-        }
-        break;
-    case ACVP_AES_CBC:
-        iv = tc->iv;
-        switch (tc->key_len) {
-        case 128:
-            cipher = EVP_aes_128_cbc();
-            break;
-        case 192:
-            cipher = EVP_aes_192_cbc();
-            break;
-        case 256:
-            cipher = EVP_aes_256_cbc();
-            break;
-        default:
-            printf("Unsupported AES key length\n");
-            return rv;
-
-            break;
-        }
-        break;
-    case ACVP_AES_XTS:
-        iv = tc->iv;
-        switch (tc->key_len) {
-        case 128:
-            cipher = EVP_aes_128_xts();
-            break;
-        case 256:
-            cipher = EVP_aes_256_xts();
-            break;
-        default:
-            printf("Unsupported AES key length\n");
-            return rv;
-
-            break;
-        }
-        break;
-    default:
-        printf("Error: Unsupported AES mode requested by ACVP server\n");
-        return rv;
-
-        break;
-    }
-
-    /* If Monte Carlo we need to be able to init and then update
-     * one thousand times before we complete each iteration.
-     */
-    if (tc->test_type == ACVP_SYM_TEST_TYPE_MCT) {
-        if (tc->direction == ACVP_SYM_CIPH_DIR_ENCRYPT) {
-            if (tc->mct_index == 0) {
-                EVP_EncryptInit_ex(cipher_ctx, cipher, NULL, tc->key, iv);
-                EVP_CIPHER_CTX_set_padding(cipher_ctx, 0);
-                if (tc->cipher == ACVP_AES_CFB1) {
-                    EVP_CIPHER_CTX_set_flags(cipher_ctx, EVP_CIPH_FLAG_LENGTH_BITS);
-                }
-            }
-            EVP_EncryptUpdate(cipher_ctx, tc->ct, &ct_len, tc->pt, tc->pt_len);
-            tc->ct_len = ct_len;
-        } else if (tc->direction == ACVP_SYM_CIPH_DIR_DECRYPT) {
-            if (tc->mct_index == 0) {
-                EVP_DecryptInit_ex(cipher_ctx, cipher, NULL, tc->key, iv);
-                EVP_CIPHER_CTX_set_padding(cipher_ctx, 0);
-                if (tc->cipher == ACVP_AES_CFB1) {
-                    EVP_CIPHER_CTX_set_flags(cipher_ctx, EVP_CIPH_FLAG_LENGTH_BITS);
-                }
-            }
-            EVP_DecryptUpdate(cipher_ctx, tc->pt, &pt_len, tc->ct, tc->ct_len);
-            tc->pt_len = pt_len;
-        } else {
-            printf("Unsupported direction\n");
-            return rv;
-        }
-        if (tc->mct_index == 999) {
-            EVP_CIPHER_CTX_cleanup(cipher_ctx);
-        }
-    } else {
-        if (tc->direction == ACVP_SYM_CIPH_DIR_ENCRYPT) {
-            EVP_EncryptInit_ex(cipher_ctx, cipher, NULL, tc->key, iv);
-            EVP_CIPHER_CTX_set_padding(cipher_ctx, 0);
-            if (tc->cipher == ACVP_AES_CFB1) {
-                EVP_CIPHER_CTX_set_flags(cipher_ctx, EVP_CIPH_FLAG_LENGTH_BITS);
-            }
-            EVP_EncryptUpdate(cipher_ctx, tc->ct, &ct_len, tc->pt, tc->pt_len);
-            tc->ct_len = ct_len;
-            EVP_EncryptFinal_ex(cipher_ctx, tc->ct + ct_len, &ct_len);
-            tc->ct_len += ct_len;
-        } else if (tc->direction == ACVP_SYM_CIPH_DIR_DECRYPT) {
-            EVP_DecryptInit_ex(cipher_ctx, cipher, NULL, tc->key, iv);
-            EVP_CIPHER_CTX_set_padding(cipher_ctx, 0);
-            if (tc->cipher == ACVP_AES_CFB1) {
-                EVP_CIPHER_CTX_set_flags(cipher_ctx, EVP_CIPH_FLAG_LENGTH_BITS);
-            }
-            EVP_DecryptUpdate(cipher_ctx, tc->pt, &pt_len, tc->ct, tc->ct_len);
-            tc->pt_len = pt_len;
-            EVP_DecryptFinal_ex(cipher_ctx, tc->pt + pt_len, &pt_len);
-            tc->pt_len += pt_len;
-        } else {
-            printf("Unsupported direction\n");
-            return rv;
-        }
-        EVP_CIPHER_CTX_cleanup(cipher_ctx);
-    }
-
-    return 0;
-}
-
-/* NOTE - openssl does not support inverse option */
-static int app_aes_keywrap_handler(ACVP_TEST_CASE *test_case) {
-    ACVP_SYM_CIPHER_TC      *tc;
-    EVP_CIPHER_CTX *cipher_ctx = NULL;
-    const EVP_CIPHER        *cipher;
-    int c_len;
-    int rc = 1;
-
-    if (!test_case) {
-        return rc;
-    }
-
-    tc = test_case->tc.symmetric;
-
-    if (tc->kwcipher != ACVP_SYM_KW_CIPHER) {
-        return rc;
-    }
-
-    /* Begin encrypt code section */
-    cipher_ctx = EVP_CIPHER_CTX_new();
-    EVP_CIPHER_CTX_init(cipher_ctx);
-
-    switch (tc->cipher) {
-    case ACVP_AES_KW:
-    case ACVP_AES_KWP:
-        switch (tc->key_len) {
-        case 128:
-            cipher = EVP_aes_128_wrap();
-            break;
-        case 192:
-            cipher = EVP_aes_192_wrap();
-            break;
-        case 256:
-            cipher = EVP_aes_256_wrap();
-            break;
-        default:
-            printf("Unsupported AES keywrap key length\n");
-            goto end;
-        }
-        break;
-    default:
-        printf("Error: Unsupported AES keywrap mode requested by ACVP server\n");
-        goto end;
-    }
-
-    if (tc->direction == ACVP_SYM_CIPH_DIR_ENCRYPT) {
-        EVP_CIPHER_CTX_set_flags(cipher_ctx, EVP_CIPHER_CTX_FLAG_WRAP_ALLOW);
-        EVP_CipherInit_ex(cipher_ctx, cipher, NULL, tc->key, NULL, 1);
-        c_len = EVP_Cipher(cipher_ctx, tc->ct, tc->pt, tc->pt_len);
-        if (c_len <= 0) {
-            printf("Error: key wrap operation failed (%d)\n", c_len);
-            goto end;
-        } else {
-            tc->ct_len = c_len;
-        }
-    } else if (tc->direction == ACVP_SYM_CIPH_DIR_DECRYPT) {
-        EVP_CIPHER_CTX_set_flags(cipher_ctx, EVP_CIPHER_CTX_FLAG_WRAP_ALLOW);
-        EVP_CipherInit_ex(cipher_ctx, cipher, NULL, tc->key, NULL, 0);
-
-#ifdef OPENSSL_KWP
-        if (tc->cipher == ACVP_AES_KWP) {
-            EVP_CIPHER_CTX_set_flags(cipher_ctx, EVP_CIPHER_CTX_FLAG_UNWRAP_WITHPAD);
-        }
-#endif
-        c_len = EVP_Cipher(cipher_ctx, tc->pt, tc->ct, tc->ct_len);
-        if (c_len <= 0) {
-            goto end;
-        } else {
-            tc->pt_len = c_len;
-        }
-    } else {
-        printf("Unsupported direction\n");
-        goto end;
-    }
-    rc = 0;
-
-end:
-    /* Cleanup */
-    if (cipher_ctx) EVP_CIPHER_CTX_free(cipher_ctx);
-
-    return rc;
-}
-
-/*
- * This fuction is invoked by libacvp when an AES crypto
- * operation is needed from the crypto module being
- * validated.  This is a callback provided to libacvp when
- * acvp_enable_capability() is invoked to register the
- * AES-GCM capabilitiy with libacvp.  libacvp will in turn
- * invoke this function when it needs to process an AES-GCM
- * test case.
- */
-static int app_aes_handler_aead(ACVP_TEST_CASE *test_case) {
-    ACVP_SYM_CIPHER_TC      *tc;
-    EVP_CIPHER_CTX *cipher_ctx = NULL;
-    const EVP_CIPHER        *cipher;
-    unsigned char iv_fixed[4] = { 1, 2, 3, 4 };
-    int rc = 0;
-    int ret = 0;
-
-    if (!test_case) {
-        return 1;
-    }
-
-    tc = test_case->tc.symmetric;
-
-    if (tc->direction != ACVP_SYM_CIPH_DIR_ENCRYPT && tc->direction != ACVP_SYM_CIPH_DIR_DECRYPT) {
-        printf("Unsupported direction\n");
-        return 1;
-    }
-
-    /* Begin encrypt code section */
-    cipher_ctx = EVP_CIPHER_CTX_new();
-    EVP_CIPHER_CTX_init(cipher_ctx);
-
-    /* Validate key length and assign OpenSSL EVP cipher */
-    switch (tc->cipher) {
-    case ACVP_AES_GCM:
-        switch (tc->key_len) {
-        case 128:
-            cipher = EVP_aes_128_gcm();
-            break;
-        case 192:
-            cipher = EVP_aes_192_gcm();
-            break;
-        case 256:
-            cipher = EVP_aes_256_gcm();
-            break;
-        default:
-            printf("Unsupported AES-GCM key length\n");
-            rc = 1;
-            goto end;
-        }
-        if (tc->direction == ACVP_SYM_CIPH_DIR_ENCRYPT) {
-            EVP_CIPHER_CTX_set_flags(cipher_ctx, EVP_CIPH_FLAG_NON_FIPS_ALLOW);
-            EVP_CipherInit(cipher_ctx, cipher, NULL, NULL, 1);
-            EVP_CIPHER_CTX_ctrl(cipher_ctx, EVP_CTRL_GCM_SET_IVLEN, tc->iv_len, 0);
-            EVP_CipherInit(cipher_ctx, NULL, tc->key, NULL, 1);
-
-            EVP_CIPHER_CTX_ctrl(cipher_ctx, EVP_CTRL_GCM_SET_IV_FIXED, 4, iv_fixed);
-            if (!EVP_CIPHER_CTX_ctrl(cipher_ctx, EVP_CTRL_GCM_IV_GEN, tc->iv_len, tc->iv)) {
-                printf("acvp_aes_encrypt: iv gen error\n");
-                rc = 1;
-                goto end;
-            }
-            if (tc->aad_len) {
-                EVP_Cipher(cipher_ctx, NULL, tc->aad, tc->aad_len);
-            }
-            EVP_Cipher(cipher_ctx, tc->ct, tc->pt, tc->pt_len);
-            EVP_Cipher(cipher_ctx, NULL, NULL, 0);
-            EVP_CIPHER_CTX_ctrl(cipher_ctx, EVP_CTRL_GCM_GET_TAG, tc->tag_len, tc->tag);
-        } else if (tc->direction == ACVP_SYM_CIPH_DIR_DECRYPT) {
-            EVP_CIPHER_CTX_set_flags(cipher_ctx, EVP_CIPH_FLAG_NON_FIPS_ALLOW);
-            EVP_CipherInit_ex(cipher_ctx, cipher, NULL, tc->key, NULL, 0);
-            EVP_CIPHER_CTX_ctrl(cipher_ctx, EVP_CTRL_GCM_SET_IVLEN, tc->iv_len, 0);
-            EVP_CIPHER_CTX_ctrl(cipher_ctx, EVP_CTRL_GCM_SET_IV_FIXED, -1, tc->iv);
-            if (!EVP_CIPHER_CTX_ctrl(cipher_ctx, EVP_CTRL_GCM_IV_GEN, tc->iv_len, tc->iv)) {
-                printf("\nFailed to set IV");
-                rc = 1;
-                goto end;
-            }
-            if (tc->aad_len) {
-                /*
-                 * Set dummy tag before processing AAD.  Otherwise the AAD can
-                 * not be processed.
-                 */
-                EVP_CIPHER_CTX_ctrl(cipher_ctx, EVP_CTRL_GCM_SET_TAG, tc->tag_len, tc->tag);
-                EVP_Cipher(cipher_ctx, NULL, tc->aad, tc->aad_len);
-            }
-            /*
-             * Set the tag when decrypting
-             */
-            EVP_CIPHER_CTX_ctrl(cipher_ctx, EVP_CTRL_GCM_SET_TAG, tc->tag_len, tc->tag);
-
-            /*
-             * Decrypt the CT
-             */
-            EVP_Cipher(cipher_ctx, tc->pt, tc->ct, tc->ct_len);
-            /*
-             * Check the tag
-             */
-            ret = EVP_Cipher(cipher_ctx, NULL, NULL, 0);
-            if (ret) {
-                rc = 1;
-                goto end;
-            }
-        }
-        break;
-    case ACVP_AES_CCM:
-        switch (tc->key_len) {
-        case 128:
-            cipher = EVP_aes_128_ccm();
-            break;
-        case 192:
-            cipher = EVP_aes_192_ccm();
-            break;
-        case 256:
-            cipher = EVP_aes_256_ccm();
-            break;
-        default:
-            printf("Unsupported AES-CCM key length\n");
-            rc = 1;
-            goto end;
-        }
-        if (tc->direction == ACVP_SYM_CIPH_DIR_ENCRYPT) {
-            EVP_CipherInit(cipher_ctx, cipher, NULL, NULL, 1);
-            EVP_CIPHER_CTX_ctrl(cipher_ctx, EVP_CTRL_CCM_SET_IVLEN, tc->iv_len, 0);
-            EVP_CIPHER_CTX_ctrl(cipher_ctx, EVP_CTRL_CCM_SET_TAG, tc->tag_len, 0);
-            EVP_CipherInit(cipher_ctx, NULL, tc->key, tc->iv, 1);
-            EVP_Cipher(cipher_ctx, NULL, NULL, tc->pt_len);
-            EVP_Cipher(cipher_ctx, NULL, tc->aad, tc->aad_len);
-            EVP_Cipher(cipher_ctx, tc->ct, tc->pt, tc->pt_len);
-            EVP_CIPHER_CTX_ctrl(cipher_ctx, EVP_CTRL_CCM_GET_TAG, tc->tag_len, tc->ct + tc->ct_len);
-            tc->ct_len += tc->tag_len;
-        } else if (tc->direction == ACVP_SYM_CIPH_DIR_DECRYPT) {
-            EVP_CipherInit(cipher_ctx, cipher, NULL, NULL, 0);
-            EVP_CIPHER_CTX_ctrl(cipher_ctx, EVP_CTRL_CCM_SET_IVLEN, tc->iv_len, 0);
-            EVP_CIPHER_CTX_ctrl(cipher_ctx, EVP_CTRL_CCM_SET_TAG, tc->tag_len, tc->ct + tc->pt_len);
-            EVP_CipherInit(cipher_ctx, NULL, tc->key, tc->iv, 0);
-            EVP_Cipher(cipher_ctx, NULL, NULL, tc->pt_len);
-            EVP_Cipher(cipher_ctx, NULL, tc->aad, tc->aad_len);
-            /*
-             * Decrypt and check the tag
-             */
-            ret = EVP_Cipher(cipher_ctx, tc->pt, tc->ct, tc->ct_len);
-            if (ret < 0) {
-                rc = 1;
-                goto end;
-            }
-        }
-        break;
-    default:
-        printf("Error: Unsupported AES AEAD mode requested by ACVP server\n");
-        rc = 1;
-        goto end;
-    }
-
-end:
-    /* Cleanup */
-    if (cipher_ctx) EVP_CIPHER_CTX_free(cipher_ctx);
-
-    return rc;
-}
-
-static int app_sha_handler(ACVP_TEST_CASE *test_case) {
-    ACVP_HASH_TC    *tc;
-    const EVP_MD    *md;
-    EVP_MD_CTX *md_ctx = NULL;
-    /* assume fail */
-    int rc = 1;
-
-    if (!test_case) {
-        return 1;
-    }
-
-    tc = test_case->tc.hash;
-    if (!tc) return rc;
-
-    switch (tc->cipher) {
-    case ACVP_HASH_SHA1:
-        md = EVP_sha1();
-        break;
-    case ACVP_HASH_SHA224:
-        md = EVP_sha224();
-        break;
-    case ACVP_HASH_SHA256:
-        md = EVP_sha256();
-        break;
-    case ACVP_HASH_SHA384:
-        md = EVP_sha384();
-        break;
-    case ACVP_HASH_SHA512:
-        md = EVP_sha512();
-        break;
-    default:
-        printf("Error: Unsupported hash algorithm requested by ACVP server\n");
-        return ACVP_NO_CAP;
-
-        break;
-    }
-    
-    if (!tc->md) {
-        printf("\nCrypto module error, md memory not allocated by library\n");
-        goto end;
-    }
-    md_ctx = EVP_MD_CTX_create();
-
-    /* If Monte Carlo we need to be able to init and then update
-     * one thousand times before we complete each iteration.
-     */
-    if (tc->test_type == ACVP_HASH_TEST_TYPE_MCT) {
-        if (!tc->m1 || !tc->m2 || !tc->m3) {
-            printf("\nCrypto module error, m1, m2, or m3 missing in sha mct test case\n");
-            goto end;
-        }
-        if (!EVP_DigestInit_ex(md_ctx, md, NULL)) {
-            printf("\nCrypto module error, EVP_DigestInit_ex failed\n");
-            goto end;
-        }
-        if (!EVP_DigestUpdate(md_ctx, tc->m1, tc->msg_len)) {
-            printf("\nCrypto module error, EVP_DigestUpdate failed\n");
-            goto end;
-        }
-        if (!EVP_DigestUpdate(md_ctx, tc->m2, tc->msg_len)) {
-            printf("\nCrypto module error, EVP_DigestUpdate failed\n");
-            goto end;
-        }
-        if (!EVP_DigestUpdate(md_ctx, tc->m3, tc->msg_len)) {
-            printf("\nCrypto module error, EVP_DigestUpdate failed\n");
-            goto end;
-        }
-        if (!EVP_DigestFinal(md_ctx, tc->md, &tc->md_len)) {
-            printf("\nCrypto module error, EVP_DigestFinal failed\n");
-            goto end;
-        }
-    } else {
-        if (!tc->msg) {
-            printf("\nCrypto module error, msg missing in sha test case\n");
-            goto end;
-        }
-        if (!EVP_DigestInit_ex(md_ctx, md, NULL)) {
-            printf("\nCrypto module error, EVP_DigestInit_ex failed\n");
-            goto end;
-        }
-
-        if (!EVP_DigestUpdate(md_ctx, tc->msg, tc->msg_len)) {
-            printf("\nCrypto module error, EVP_DigestUpdate failed\n");
-            goto end;
-        }
-        if (!EVP_DigestFinal(md_ctx, tc->md, &tc->md_len)) {
-            printf("\nCrypto module error, EVP_DigestFinal failed\n");
-            goto end;
-        }
-    }
-
-    rc = 0;
-
-end:
-    if (md_ctx) EVP_MD_CTX_destroy(md_ctx);
-
-    return rc;
-}
-
-static int app_hmac_handler(ACVP_TEST_CASE *test_case) {
-    ACVP_HMAC_TC    *tc;
-    const EVP_MD    *md;
-    HMAC_CTX *hmac_ctx = NULL;
-    int msg_len;
-    int rc = 1;
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-    HMAC_CTX static_ctx;
-#endif
-
-    if (!test_case) {
-        return rc;
-    }
-
-    tc = test_case->tc.hmac;
-    if (!tc) return rc;
-    
-    switch (tc->cipher) {
-    case ACVP_HMAC_SHA1:
-        md = EVP_sha1();
-        break;
-    case ACVP_HMAC_SHA2_224:
-        md = EVP_sha224();
-        break;
-    case ACVP_HMAC_SHA2_256:
-        md = EVP_sha256();
-        break;
-    case ACVP_HMAC_SHA2_384:
-        md = EVP_sha384();
-        break;
-    case ACVP_HMAC_SHA2_512:
-        md = EVP_sha512();
-        break;
-    default:
-        printf("Error: Unsupported hash algorithm requested by ACVP server\n");
-        return rc;
-
-        break;
-    }
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-    hmac_ctx = &static_ctx;
-    HMAC_CTX_init(hmac_ctx);
-#else
-    hmac_ctx = HMAC_CTX_new();
-#endif
-    msg_len = tc->msg_len;
-
-    if (!HMAC_Init_ex(hmac_ctx, tc->key, tc->key_len, md, NULL)) {
-        printf("\nCrypto module error, HMAC_Init_ex failed\n");
-        goto end;
-    }
-
-    if (!HMAC_Update(hmac_ctx, tc->msg, msg_len)) {
-        printf("\nCrypto module error, HMAC_Update failed\n");
-        goto end;
-    }
-
-    if (!HMAC_Final(hmac_ctx, tc->mac, &tc->mac_len)) {
-        printf("\nCrypto module error, HMAC_Final failed\n");
-        goto end;
-    }
-
-    rc = 0;
-
-end:
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-    HMAC_CTX_cleanup(hmac_ctx);
-#else
-    if (hmac_ctx) HMAC_CTX_free(hmac_ctx);
-#endif
-
-    return rc;
-}
-
-static int app_cmac_handler(ACVP_TEST_CASE *test_case) {
-    ACVP_CMAC_TC    *tc;
-    int rv = 1;
-    const EVP_CIPHER    *c = NULL;
-    CMAC_CTX       *cmac_ctx = NULL;
-    int key_len, i;
-    unsigned char mac_compare[16] = { 0 };
-    char full_key[32] = { 0 };
-    int mac_cmp_len;
-
-    if (!test_case) {
-        return rv;
-    }
-
-    tc = test_case->tc.cmac;
-    if (!tc) return rv;
-    if (!tc->key) return rv;
-
-    switch (tc->cipher) {
-    case ACVP_CMAC_AES:
-        switch (tc->key_len * 8) {
-        case 128:
-            c = EVP_aes_128_cbc();
-            break;
-        case 192:
-            c = EVP_aes_192_cbc();
-            break;
-        case 256:
-            c = EVP_aes_256_cbc();
-            break;
-        default:
-            break;
-        }
-        key_len = (tc->key_len);
-        for (i = 0; i < key_len; i++) {
-            full_key[i] = tc->key[i];
-        }
-        break;
-    case ACVP_CMAC_TDES:
-        c = EVP_des_ede3_cbc();
-        for (i = 0; i < 8; i++) {
-            full_key[i] = tc->key[i];
-        }
-        for (; i < 16; i++) {
-            full_key[i] = tc->key2[i % 8];
-        }
-        for (; i < 24; i++) {
-            full_key[i] = tc->key3[i % 8];
-        }
-        key_len = 24;
-        break;
-    default:
-        printf("Error: Unsupported CMAC algorithm requested by ACVP server\n");
-        return rv;
-    }
-
-    full_key[key_len] = '\0';
-
-    cmac_ctx = CMAC_CTX_new();
-
-    if (!CMAC_Init(cmac_ctx, full_key, key_len, c, NULL)) {
-        printf("\nCrypto module error, CMAC_Init_ex failed\n");
-        goto cleanup;
-    }
-
-    if (!CMAC_Update(cmac_ctx, tc->msg, tc->msg_len)) {
-        printf("\nCrypto module error, CMAC_Update failed\n");
-        goto cleanup;
-    }
-
-    if (tc->verify) {
-        int diff = 0;
-
-        if (!CMAC_Final(cmac_ctx, mac_compare, (size_t *)&mac_cmp_len)) {
-            printf("\nCrypto module error, CMAC_Final failed\n");
-            goto cleanup;
-        }
-
-        memcmp_s(tc->mac, tc->mac_len, mac_compare, mac_cmp_len, &diff);
-        if (!diff) {
-            tc->ver_disposition = ACVP_TEST_DISPOSITION_PASS;
-        } else {
-            tc->ver_disposition = ACVP_TEST_DISPOSITION_FAIL;
-        }
-    } else {
-        if (!CMAC_Final(cmac_ctx, tc->mac, (size_t *)&tc->mac_len)) {
-            printf("\nCrypto module error, CMAC_Final failed\n");
-            goto cleanup;
-        }
-    }
-    rv = 0;
-
-cleanup:
-    if (cmac_ctx) CMAC_CTX_free(cmac_ctx);
-
-    return rv;
-}
-
-#ifdef OPENSSL_KDF_SUPPORT
-static int app_kdf135_srtp_handler(ACVP_TEST_CASE *test_case) {
-    return 1;
-}
-
-static int app_kdf135_ikev2_handler(ACVP_TEST_CASE *test_case) {
-    return 1;
-}
-
-static int app_kdf135_ikev1_handler(ACVP_TEST_CASE *test_case) {
-    return 1;
-}
-
-static int app_kdf135_x963_handler(ACVP_TEST_CASE *test_case) {
-    return 1;
-}
-
-static int app_kdf108_handler(ACVP_TEST_CASE *test_case) {
-    return 1;
-}
-
-static int app_kdf135_tls_handler(ACVP_TEST_CASE *test_case) {
-    return 1; 
-}
-
-static int app_kdf135_snmp_handler(ACVP_TEST_CASE *test_case) {
-    return 1;
-}
-
-static int app_kdf135_ssh_handler(ACVP_TEST_CASE *test_case) {
-    return 1;
-}
-#endif
-
-//Must be commented out if the user is Making with Makefile.fom
-#ifdef ACVP_NO_RUNTIME
-/*
- * group values are declared near the top so that they
- * can be freed at the end of main execution
- */
-static int app_dsa_handler(ACVP_TEST_CASE *test_case) {
-    int L, N, n, r;
-    const EVP_MD        *md = NULL;
-    ACVP_DSA_TC         *tc;
-    unsigned char seed[1024];
-    DSA                 *dsa = NULL;
-    int counter, counter2;
-    unsigned long h, h2;
-    DSA_SIG             *sig = NULL;
-    BIGNUM              *q = NULL, *p = NULL, *g = NULL;
-    BIGNUM              *q2 = NULL, *p2 = NULL, *g2 = NULL;
-    BIGNUM *priv_key = NULL, *pub_key = NULL;
-    BIGNUM *sig_r = NULL, *sig_s = NULL;
-
-
-    tc = test_case->tc.dsa;
-    switch (tc->mode) {
-    case ACVP_DSA_MODE_KEYGEN:
-        if (dsa_current_keygen_tg != tc->tg_id) {
-            dsa_current_keygen_tg = tc->tg_id;
-
-            if (group_dsa) FIPS_dsa_free(group_dsa);
-            if (group_p) BN_free(group_p);
-            if (group_q) BN_free(group_q);
-            if (group_g) BN_free(group_g);
-            if (group_pub_key) BN_free(group_pub_key);
-
-            group_dsa = FIPS_dsa_new();
-            if (!group_dsa) {
-                printf("Failed to allocate DSA strcut\n");
-                return 1;
-            }
-            L = tc->l;
-            N = tc->n;
-
-            if (dsa_builtin_paramgen2(group_dsa, L, N, md, NULL, 0, -1,
-                                      NULL, NULL, NULL, NULL) <= 0) {
-                printf("Parameter Generation error\n");
-                FIPS_dsa_free(group_dsa);
-                return 1;
-            }
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-            group_p = BN_dup(group_dsa->p);
-            group_q = BN_dup(group_dsa->q);
-            group_g = BN_dup(group_dsa->g);
-#else
-            DSA_get0_pqg(group_dsa, (const BIGNUM **)&group_p,
-                         (const BIGNUM **)&group_q, (const BIGNUM **)&group_g);
-#endif
-        }
-
-        tc->p_len = BN_bn2bin(group_p, tc->p);
-        tc->q_len = BN_bn2bin(group_q, tc->q);
-        tc->g_len = BN_bn2bin(group_g, tc->g);
-
-        if (!DSA_generate_key(group_dsa)) {
-            printf("\n DSA_generate_key failed");
-            FIPS_dsa_free(group_dsa);
-            return 1;
-        }
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-        priv_key = group_dsa->priv_key;
-        pub_key = group_dsa->pub_key;
-#else
-        DSA_get0_key(group_dsa, (const BIGNUM **)&pub_key,
-                     (const BIGNUM **)&priv_key);
-#endif
-
-        tc->x_len = BN_bn2bin(priv_key, tc->x);
-        tc->y_len = BN_bn2bin(pub_key, tc->y);
-        break;
-
-    case ACVP_DSA_MODE_PQGVER:
-        switch (tc->sha) {
-        case ACVP_SHA1:
-            md = EVP_sha1();
-            break;
-        case ACVP_SHA224:
-            md = EVP_sha224();
-            break;
-        case ACVP_SHA256:
-            md = EVP_sha256();
-            break;
-        case ACVP_SHA384:
-            md = EVP_sha384();
-            break;
-        case ACVP_SHA512:
-            md = EVP_sha512();
-            break;
-        case ACVP_SHA512_224:
-        case ACVP_SHA512_256:
-        default:
-            printf("DSA sha value not supported %d\n", tc->sha);
-            return 1;
-
-            break;
-        }
-
-        switch (tc->pqg) {
-        case ACVP_DSA_PROBABLE:
-            dsa = FIPS_dsa_new();
-            if (!dsa) {
-                printf("Failed to allocate DSA strcut\n");
-                return 1;
-            }
-            L = tc->l;
-            N = tc->n;
-
-            p = FIPS_bn_new();
-            q = FIPS_bn_new();
-            BN_bin2bn(tc->p, tc->p_len, p);
-            BN_bin2bn(tc->q, tc->q_len, q);
-
-            if (dsa_builtin_paramgen2(dsa, L, N, md,
-                                      tc->seed, tc->seedlen, -1, NULL,
-                                      &counter2, &h2, NULL) < 0) {
-                printf("Parameter Generation error\n");
-                FIPS_dsa_free(dsa);
-                return 1;
-            }
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-            p2 = BN_dup(dsa->p);
-            q2 = BN_dup(dsa->q);
-#else
-            DSA_get0_pqg(dsa, (const BIGNUM **)&p2,
-                         (const BIGNUM **)&q2, NULL);
-#endif
-
-            if (BN_cmp(p2, p) || BN_cmp(q2, q))
-                r = 0;
-            else
-                r = 1;
-
-            FIPS_dsa_free(dsa);
-            tc->result = r;
-            break;
-
-        case ACVP_DSA_CANONICAL:
-            dsa = FIPS_dsa_new();
-            if (!dsa) {
-                printf("Failed to allocate DSA strcut\n");
-                return 1;
-            }
-            L = tc->l;
-            N = tc->n;
-
-            p = FIPS_bn_new();
-            q = FIPS_bn_new();
-            g = FIPS_bn_new();
-            BN_bin2bn(tc->p, tc->p_len, p);
-            BN_bin2bn(tc->q, tc->q_len, q);
-            BN_bin2bn(tc->g, tc->g_len, g);
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-            dsa->p = BN_dup(p);
-            dsa->q = BN_dup(q);
-#else
-            DSA_set0_pqg(dsa, BN_dup(p), BN_dup(q), NULL);
-#endif
-
-            if (dsa_builtin_paramgen2(dsa, L, N, md,
-                                      tc->seed, tc->seedlen, tc->index, NULL,
-                                      &counter2, &h2, NULL) < 0) {
-                printf("Parameter Generation error\n");
-                FIPS_dsa_free(dsa);
-                return 1;
-            }
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-            g2 = BN_dup(dsa->g);
-#else
-            DSA_get0_pqg(dsa, NULL, NULL, (const BIGNUM **)&g2);
-#endif
-
-            if (BN_cmp(g2, g)) {
-                r = 0;
-            } else {
-                r = 1;
-            }
-            FIPS_dsa_free(dsa);
-            tc->result = r;
-            break;
-        default:
-            printf("DSA pqg mode not supported %d\n", tc->pqg);
-            return 1;
-
-            break;
-        }
-        break;
-
-    case ACVP_DSA_MODE_SIGVER:
-        switch (tc->sha) {
-        case ACVP_SHA1:
-            md = EVP_sha1();
-            break;
-        case ACVP_SHA224:
-            md = EVP_sha224();
-            break;
-        case ACVP_SHA256:
-            md = EVP_sha256();
-            break;
-        case ACVP_SHA384:
-            md = EVP_sha384();
-            break;
-        case ACVP_SHA512:
-            md = EVP_sha512();
-            break;
-        case ACVP_SHA512_224:
-        case ACVP_SHA512_256:
-        default:
-            printf("DSA sha value not supported %d\n", tc->sha);
-            return 1;
-
-            break;
-        }
-
-        dsa = FIPS_dsa_new();
-        if (!dsa) {
-            printf("Failed to allocate DSA strcut\n");
-            return 1;
-        }
-        sig = FIPS_dsa_sig_new();
-        if (!sig) {
-            printf("Failed to allocate SIG strcut\n");
-            FIPS_dsa_free(dsa);
-            return 1;
-        }
-        L = tc->l;
-        N = tc->n;
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-        dsa->p = BN_new();
-        dsa->q = BN_new();
-        dsa->g = BN_new();
-        dsa->pub_key = BN_new();
-        sig->r = BN_new();
-        sig->s = BN_new();
-        BN_bin2bn(tc->p, tc->p_len, dsa->p);
-        BN_bin2bn(tc->q, tc->q_len, dsa->q);
-        BN_bin2bn(tc->g, tc->g_len, dsa->g);
-        BN_bin2bn(tc->y, tc->y_len, dsa->pub_key);
-        BN_bin2bn(tc->r, tc->r_len, sig->r);
-        BN_bin2bn(tc->s, tc->s_len, sig->s);
-#else
-        DSA_get0_pqg(dsa, (const BIGNUM **)&p,
-                     (const BIGNUM **)&q, (const BIGNUM **)&g);
-        DSA_get0_key(dsa, (const BIGNUM **)&pub_key, NULL);
-        DSA_SIG_get0(sig, (const BIGNUM **)&sig_r, (const BIGNUM **)&sig_s);
-
-        BN_bin2bn(tc->p, tc->p_len, p);
-        BN_bin2bn(tc->q, tc->q_len, q);
-        BN_bin2bn(tc->g, tc->g_len, g);
-        BN_bin2bn(tc->y, tc->y_len, pub_key);
-        BN_bin2bn(tc->r, tc->r_len, sig_r);
-        BN_bin2bn(tc->s, tc->s_len, sig_s);
-#endif
-
-        n = tc->msglen;
-        r = FIPS_dsa_verify(dsa, (const unsigned char *)tc->msg, n, md, sig);
-
-        FIPS_dsa_free(dsa);
-        FIPS_dsa_sig_free(sig);
-        /* return result, 0 is failure, 1 is pass */
-        tc->result = r;
-        break;
-
-    case ACVP_DSA_MODE_SIGGEN:
-        switch (tc->sha) {
-        case ACVP_SHA1:
-            md = EVP_sha1();
-            break;
-        case ACVP_SHA224:
-            md = EVP_sha224();
-            break;
-        case ACVP_SHA256:
-            md = EVP_sha256();
-            break;
-        case ACVP_SHA384:
-            md = EVP_sha384();
-            break;
-        case ACVP_SHA512:
-            md = EVP_sha512();
-            break;
-        case ACVP_SHA512_224:
-        case ACVP_SHA512_256:
-        default:
-            printf("DSA sha value not supported %d\n", tc->sha);
-            return 1;
-
-            break;
-        }
-
-        if (dsa_current_siggen_tg != tc->tg_id) {
-            dsa_current_siggen_tg = tc->tg_id;
-
-            if (group_dsa) FIPS_dsa_free(group_dsa);
-            if (group_p) BN_free(group_p);
-            if (group_q) BN_free(group_q);
-            if (group_g) BN_free(group_g);
-            if (group_pub_key) BN_free(group_pub_key);
-
-            group_dsa = FIPS_dsa_new();
-            if (!group_dsa) {
-                printf("Failed to allocate DSA strcut\n");
-                return 1;
-            }
-            L = tc->l;
-            N = tc->n;
-
-            if (dsa_builtin_paramgen2(group_dsa, L, N, md, NULL, 0, -1,
-                                      NULL, NULL, NULL, NULL) <= 0) {
-                printf("Parameter Generation error\n");
-                FIPS_dsa_free(group_dsa);
-                return 1;
-            }
-
-            if (!DSA_generate_key(group_dsa)) {
-                printf("\n DSA_generate_key failed");
-                FIPS_dsa_free(group_dsa);
-                return 1;
-            }
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-            group_p = BN_dup(group_dsa->p);
-            group_q = BN_dup(group_dsa->q);
-            group_g = BN_dup(group_dsa->g);
-#else
-            DSA_get0_pqg(group_dsa, (const BIGNUM **)&group_p,
-                         (const BIGNUM **)&group_q, (const BIGNUM **)&group_g);
-#endif
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-            group_pub_key = BN_dup(group_dsa->pub_key);
-#else
-            DSA_get0_key(group_dsa, (const BIGNUM **)&group_pub_key, NULL);
-#endif
-        }
-
-        tc->p_len = BN_bn2bin(group_p, tc->p);
-        tc->q_len = BN_bn2bin(group_q, tc->q);
-        tc->g_len = BN_bn2bin(group_g, tc->g);
-        tc->y_len = BN_bn2bin(group_pub_key, tc->y);
-
-        sig = FIPS_dsa_sign(group_dsa, tc->msg, tc->msglen, md);
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-        sig_r = sig->r;
-        sig_s = sig->s;
-#else
-        DSA_SIG_get0(sig, (const BIGNUM **)&sig_r, (const BIGNUM **)&sig_s);
-#endif
-
-        tc->r_len = BN_bn2bin(sig_r, tc->r);
-        tc->s_len = BN_bn2bin(sig_s, tc->s);
-        FIPS_dsa_sig_free(sig);
-        break;
-
-    case ACVP_DSA_MODE_PQGGEN:
-        switch (tc->sha) {
-        case ACVP_SHA1:
-            md = EVP_sha1();
-            break;
-        case ACVP_SHA224:
-            md = EVP_sha224();
-            break;
-        case ACVP_SHA256:
-            md = EVP_sha256();
-            break;
-        case ACVP_SHA384:
-            md = EVP_sha384();
-            break;
-        case ACVP_SHA512:
-            md = EVP_sha512();
-            break;
-        case ACVP_SHA512_224:
-        case ACVP_SHA512_256:
-        default:
-            printf("DSA sha value not supported %d\n", tc->sha);
-            return 1;
-
-            break;
-        }
-
-        switch (tc->gen_pq) {
-        case ACVP_DSA_UNVERIFIABLE:
-            printf("DSA Parameter Generation2 error for %d, not supported\n", tc->gen_pq);
-            return 1;
-
-            break;
-        case ACVP_DSA_CANONICAL:
-            dsa = FIPS_dsa_new();
-
-            p = FIPS_bn_new();
-            q = FIPS_bn_new();
-            BN_bin2bn(tc->p, tc->p_len, p);
-            BN_bin2bn(tc->q, tc->q_len, q);
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-            dsa->p = BN_dup(p);
-            dsa->q = BN_dup(q);
-#else
-            DSA_set0_pqg(dsa, p, q, g);
-#endif
-            L = tc->l;
-            N = tc->n;
-            if (dsa_builtin_paramgen2(dsa, L, N, md,
-                                      tc->seed, tc->seedlen, tc->index, NULL,
-                                      NULL, NULL, NULL) <= 0) {
-                printf("DSA Parameter Generation2 error for %d\n", tc->gen_pq);
-                FIPS_dsa_free(dsa);
-                return 1;
-            }
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-            tc->g_len = BN_bn2bin(dsa->g, tc->g);
-#else
-            tc->g_len = BN_bn2bin(g, tc->g);
-#endif
-            FIPS_dsa_free(dsa);
-            break;
-
-        case ACVP_DSA_PROBABLE:
-        case ACVP_DSA_PROVABLE:
-            dsa = FIPS_dsa_new();
-            L = tc->l;
-            N = tc->n;
-            if (dsa_builtin_paramgen2(dsa, L, N, md,
-                                      NULL, 0, -1, seed,
-                                      &counter, &h, NULL) <= 0) {
-                printf("DSA Parameter Generation 2 error for %d\n", tc->gen_pq);
-                FIPS_dsa_free(dsa);
-                return 1;
-            }
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-            p = dsa->p;
-            q = dsa->q;
-#else
-            DSA_get0_pqg(dsa, (const BIGNUM **)&p,
-                         (const BIGNUM **)&q, NULL);
-#endif
-
-            tc->p_len = BN_bn2bin(p, tc->p);
-            tc->q_len = BN_bn2bin(q, tc->q);
-            tc->counter = counter;
-            tc->h = h;
-
-#define DSA_MAX_SEED 3072
-            memcpy_s(tc->seed, DSA_MAX_SEED, &seed, EVP_MD_size(md));
-            tc->seedlen = EVP_MD_size(md);
-            tc->counter = counter;
-            FIPS_dsa_free(dsa);
-            break;
-        default:
-            printf("Invalid DSA gen_pq %d\n", tc->gen_pq);
-            return 1;
-
-            break;
-        }
-        break;
-    default:
-        printf("Invalid DSA mode %d\n", tc->mode);
-        return 1;
-
-        break;
-    }
-    return 0;
-}
-
-static EC_POINT *make_peer(EC_GROUP *group, BIGNUM *x, BIGNUM *y) {
-    EC_POINT *peer;
-    int rv;
-    BN_CTX *c;
-
-    peer = EC_POINT_new(group);
-    if (!peer) {
-        printf("EC_POINT_new failed\n");
-        return NULL;
-    }
-    c = BN_CTX_new();
-    if (!c) {
-        printf("BN_CTX_new failed\n");
-        return NULL;
-    }
-    if (EC_METHOD_get_field_type(EC_GROUP_method_of(group))
-        == NID_X9_62_prime_field) {
-        rv = EC_POINT_set_affine_coordinates_GFp(group, peer, x, y, c);
-    } else {
-        rv = EC_POINT_set_affine_coordinates_GF2m(group, peer, x, y, c);
-    }
-
-    BN_CTX_free(c);
-    if (rv) {
-        return peer;
-    }
-    EC_POINT_free(peer);
-    return NULL;
-}
-
-static int ec_print_key(ACVP_KAS_ECC_TC *tc, EC_KEY *key, int add_e, int exout) {
-    const EC_POINT *pt;
-    const EC_GROUP *grp;
-    const EC_METHOD *meth;
-    int rv = 0;
-    BIGNUM *tx, *ty;
-    const BIGNUM *d = NULL;
-    BN_CTX *ctx;
-
-    ctx = BN_CTX_new();
-    if (!ctx) {
-        printf("BN_CTX_new failed\n");
-        return 0;
-    }
-    tx = BN_CTX_get(ctx);
-    ty = BN_CTX_get(ctx);
-    if (!tx || !ty) {
-        BN_CTX_free(ctx);
-        printf("BN_CTX_get failed\n");
-        return 0;
-    }
-    grp = EC_KEY_get0_group(key);
-    pt = EC_KEY_get0_public_key(key);
-    d = EC_KEY_get0_private_key(key);
-    meth = EC_GROUP_method_of(grp);
-    if (EC_METHOD_get_field_type(meth) == NID_X9_62_prime_field) {
-        rv = EC_POINT_get_affine_coordinates_GFp(grp, pt, tx, ty, ctx);
-    } else {
-        rv = EC_POINT_get_affine_coordinates_GF2m(grp, pt, tx, ty, ctx);
-    }
-
-    if (tc->test_type == ACVP_KAS_ECC_TT_AFT) {
-        tc->pixlen = BN_bn2bin(tx, tc->pix);
-        tc->piylen = BN_bn2bin(ty, tc->piy);
-        if (tc->mode == ACVP_KAS_ECC_MODE_COMPONENT) {
-            tc->dlen = BN_bn2bin(d, tc->d);
-        }
-    }
-    BN_CTX_free(ctx);
-    return rv;
-}
-
-static int app_kas_ecc_handler(ACVP_TEST_CASE *test_case) {
-    EC_GROUP *group = NULL;
-    ACVP_KAS_ECC_TC         *tc;
-    int nid = 0, exout = 0;
-    EC_KEY *ec = NULL;
-    EC_POINT *peerkey = NULL;
-    unsigned char *Z = NULL;
-    int Zlen = 0;
-    BIGNUM *cx = NULL, *cy = NULL, *ix = NULL, *iy = NULL, *id = NULL;
-    const EVP_MD *md = NULL;
-    int rv = 1;
-
-    tc = test_case->tc.kas_ecc;
-
-    switch (tc->curve) {
-    case ACVP_EC_CURVE_P224:
-        nid = NID_secp224r1;
-        break;
-    case ACVP_EC_CURVE_P256:
-        nid = NID_X9_62_prime256v1;
-        break;
-    case ACVP_EC_CURVE_P384:
-        nid = NID_secp384r1;
-        break;
-    case ACVP_EC_CURVE_P521:
-        nid = NID_secp521r1;
-        break;
-    case ACVP_EC_CURVE_B233:
-        nid = NID_sect233r1;
-        break;
-    case ACVP_EC_CURVE_B283:
-        nid = NID_sect283r1;
-        break;
-    case ACVP_EC_CURVE_B409:
-        nid = NID_sect409r1;
-        break;
-    case ACVP_EC_CURVE_B571:
-        nid = NID_sect571r1;
-        break;
-    case ACVP_EC_CURVE_K233:
-        nid = NID_sect233k1;
-        break;
-    case ACVP_EC_CURVE_K283:
-        nid = NID_sect283k1;
-        break;
-    case ACVP_EC_CURVE_K409:
-        nid = NID_sect409k1;
-        break;
-    case ACVP_EC_CURVE_K571:
-        nid = NID_sect571k1;
-        break;
-    default:
-        printf("Invalid curve %d\n", tc->curve);
-        return rv;
-
-        break;
-    }
-
-    if (tc->mode == ACVP_KAS_ECC_MODE_COMPONENT) {
-        switch (tc->md) {
-        case ACVP_SHA224:
-            md = EVP_sha224();
-            break;
-        case ACVP_SHA256:
-            md = EVP_sha256();
-            break;
-        case ACVP_SHA384:
-            md = EVP_sha384();
-            break;
-        case ACVP_SHA512:
-            md = EVP_sha512();
-            break;
-        default:
-            printf("No valid hash name %d\n", tc->md);
-            return rv;
-
-            break;
-        }
-    }
-    group = EC_GROUP_new_by_curve_name(nid);
-    if (group == NULL) {
-        printf("No group from curve name %d\n", nid);
-        return rv;
-    }
-
-    ec = EC_KEY_new();
-    if (ec == NULL) {
-        EC_GROUP_free(group);
-        printf("No EC_KEY_new\n");
-        return rv;
-    }
-    EC_KEY_set_flags(ec, EC_FLAG_COFACTOR_ECDH);
-    if (!EC_KEY_set_group(ec, group)) {
-        EC_GROUP_free(group);
-        printf("No EC_KEY_set_group\n");
-        return rv;
-    }
-
-    cx = FIPS_bn_new();
-    cy = FIPS_bn_new();
-    BN_bin2bn(tc->psx, tc->psxlen, cx);
-    BN_bin2bn(tc->psy, tc->psylen, cy);
-    if (!cx || !cy) {
-        printf("BN_bin2bn failed psx psy\n");
-        goto error;
-    }
-    peerkey = make_peer(group, cx, cy);
-    if (peerkey == NULL) {
-        printf("Peerkey failed\n");
-        goto error;
-    }
-    if (tc->test_type == ACVP_KAS_ECC_TT_VAL) {
-        ix = FIPS_bn_new();
-        iy = FIPS_bn_new();
-        id = FIPS_bn_new();
-        BN_bin2bn(tc->pix, tc->pixlen, ix);
-        BN_bin2bn(tc->piy, tc->piylen, iy);
-        BN_bin2bn(tc->d, tc->dlen, id);
-
-        if (!ix || !iy || !id) {
-            printf("BN_bin2bn failed pix piy d");
-            goto error;
-        }
-
-        EC_KEY_set_public_key_affine_coordinates(ec, ix, iy);
-        EC_KEY_set_private_key(ec, id);
-    } else {
-        if (!EC_KEY_generate_key(ec)) {
-            printf("EC_KEY_generate_key failed\n");
-            goto error;
-        }
-    }
-
-    exout = md ? 1 : 0;
-    ec_print_key(tc, ec, md ? 1 : 0, exout);
-    Zlen = (EC_GROUP_get_degree(group) + 7) / 8;
-    if (!Zlen) {
-        printf("Zlen degree failure\n");
-        goto error;
-    }
-    Z = OPENSSL_malloc(Zlen);
-    if (!Z) {
-        printf("Malloc failure\n");
-        goto error;
-    }
-    if (!ECDH_compute_key(Z, Zlen, peerkey, ec, 0)) {
-        printf("ECDH_compute_key failure\n");
-        goto error;
-    }
-
-#define KAS_ECC_Z_MAX 512
-    if (tc->test_type == ACVP_KAS_ECC_TT_AFT) {
-        memcpy_s(tc->z, KAS_ECC_Z_MAX, Z, Zlen);
-        tc->zlen = Zlen;
-    }
-    if (tc->mode == ACVP_KAS_ECC_MODE_COMPONENT) {
-        FIPS_digest(Z, Zlen, (unsigned char *)tc->chash, NULL, md);
-        tc->chashlen = EVP_MD_size(md);
-    }
-    rv = 0;
-
-error:
-    if (Z) {
-        OPENSSL_cleanse(Z, Zlen);
-    }
-    FIPS_free(Z);
-    EC_KEY_free(ec);
-    EC_POINT_free(peerkey);
-    EC_GROUP_free(group);
-    BN_free(cx);
-    BN_free(cy);
-    BN_free(ix);
-    BN_free(iy);
-    BN_free(id);
-    return rv;
-}
-
-static int app_kas_ffc_handler(ACVP_TEST_CASE *test_case) {
-    ACVP_KAS_FFC_TC         *tc;
-    const EVP_MD *md = NULL;
-    int rv = 1;
-    unsigned char *Z = NULL;
-    int Zlen = 0;
-    DH *dh = NULL;
-    BIGNUM *p = NULL, *q = NULL, *g = NULL;
-    BIGNUM *pub_key = NULL, *priv_key = NULL;
-    BIGNUM *peerkey = NULL;
-
-    tc = test_case->tc.kas_ffc;
-
-    switch (tc->md) {
-    case ACVP_SHA224:
-        md = EVP_sha224();
-        break;
-    case ACVP_SHA256:
-        md = EVP_sha256();
-        break;
-    case ACVP_SHA384:
-        md = EVP_sha384();
-        break;
-    case ACVP_SHA512:
-        md = EVP_sha512();
-        break;
-    default:
-        printf("No valid hash name %d\n", tc->md);
-        return rv;
-
-        break;
-    }
-
-    dh = FIPS_dh_new();
-    if (!dh) {
-        return rv;
-    }
-
-    p = FIPS_bn_new();
-    q = FIPS_bn_new();
-    g = FIPS_bn_new();
-    BN_bin2bn(tc->p, tc->plen, p);
-    BN_bin2bn(tc->q, tc->qlen, q);
-    BN_bin2bn(tc->g, tc->glen, g);
-
-    peerkey = FIPS_bn_new();
-    BN_bin2bn(tc->eps, tc->epslen, peerkey);
-
-    if (!peerkey || !p || !q || !g) {
-        printf("BN_bin2bn failed p q g eps\n");
-        goto error;
-    }
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-    dh->p = BN_dup(p);
-    dh->q = BN_dup(q);
-    dh->g = BN_dup(g);
-#else
-    DH_set0_pqg(dh, p, q, g);
-#endif
-
-    if (tc->test_type == ACVP_KAS_FFC_TT_VAL) {
-        pub_key = FIPS_bn_new();
-        priv_key = FIPS_bn_new();
-        BN_bin2bn(tc->epri, tc->eprilen, priv_key);
-        BN_bin2bn(tc->epui, tc->epuilen, pub_key);
-
-        if (!pub_key || !priv_key) {
-            printf("BN_bin2bn failed epri epui\n");
-            goto error;
-        }
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-        dh->pub_key = BN_dup(pub_key);
-        dh->priv_key = BN_dup(priv_key);
-#else
-        DH_set0_key(dh, pub_key, priv_key);
-#endif
-    }
-
-    if (tc->test_type == ACVP_KAS_FFC_TT_AFT) {
-        if (!DH_generate_key(dh)) {
-            printf("DH_generate_key failed\n");
-            goto error;
-        }
-    }
-    Z = OPENSSL_malloc(BN_num_bytes(p));
-    if (!Z) {
-        printf("Malloc failed for Z\n");
-        goto error;
-    }
-
-    Zlen = DH_compute_key_padded(Z, peerkey, dh);
-    FIPS_digest(Z, Zlen, (unsigned char *)tc->chash, NULL, md);
-    tc->chashlen = EVP_MD_size(md);
-
-#define KAS_FFC_Z_MAX 512
-    if (tc->test_type == ACVP_KAS_FFC_TT_AFT) {
-        memcpy_s(tc->z, KAS_FFC_Z_MAX, Z, Zlen);
-        tc->zlen = Zlen;
-    }
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-    tc->piutlen = BN_bn2bin(dh->pub_key, tc->piut);
-#else
-    tc->piutlen = BN_bn2bin(pub_key, tc->piut);
-#endif
-
-    rv = 0;
-
-error:
-    if (Z) {
-        OPENSSL_cleanse(Z, Zlen);
-    }
-    FIPS_free(Z);
-    BN_clear_free(peerkey);
-    FIPS_dh_free(dh);
-    return rv;
-}
-
-static int app_rsa_keygen_handler(ACVP_TEST_CASE *test_case) {
-    /*
-     * custom crypto module handler
-     * to be filled in -
-     * this handler assumes info gen by server
-     * and all the other params registered for
-     * in this example app.
-     */
-
-    ACVP_RSA_KEYGEN_TC    *tc;
-    int rv = 1;
-    RSA       *rsa;
-    BIGNUM *p = NULL, *q = NULL, *n = NULL, *d = NULL;
-    BIGNUM *e = BN_new();
-
-    if (!test_case) {
-        goto err;
-    }
-    tc = test_case->tc.rsa_keygen;
-
-    rsa = FIPS_rsa_new();
-
-    e = FIPS_bn_new();
-    BN_bin2bn(tc->e, tc->e_len, e);
-    if (!e) {
-        printf("Error converting e to BN\n");
-        goto err;
-    }
-
-    /*
-     * IMPORTANT: Placeholder! The RSA keygen vector
-     * sets will fail if this handler is left as is.
-     *
-     * Below, insert your own key generation API that
-     * supports specification of all of the params...
-     */
-    if (!FIPS_rsa_x931_generate_key_ex(rsa, tc->modulo, e, NULL)) {
-        printf("\nError: Issue with key generation\n");
-        goto err;
-    }
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-    p = rsa->p;
-    q = rsa->q;
-    n = rsa->n;
-    d = rsa->d;
-#else
-    RSA_get0_key(rsa, (const BIGNUM **)&n, NULL,
-                 (const BIGNUM **)&d);
-    RSA_get0_factors(rsa, (const BIGNUM **)&p,
-                     (const BIGNUM **)&q);
-#endif
-
-    tc->p_len = BN_bn2bin(p, tc->p);
-    tc->q_len = BN_bn2bin(q, tc->q);
-    tc->n_len = BN_bn2bin(n, tc->n);
-    tc->d_len = BN_bn2bin(d, tc->d);
-
-    FIPS_rsa_free(rsa);
-    rv = 0;
-err:
-    if (e) BN_free(e);
-    return rv;
-}
-
-static int ec_get_pubkey(EC_KEY *key, BIGNUM *x, BIGNUM *y) {
-    const EC_POINT *pt;
-    const EC_GROUP *grp;
-    const EC_METHOD *meth;
-    int rv;
-    BN_CTX *ctx;
-
-    ctx = BN_CTX_new();
-    if (!ctx)
-        return 0;
-
-    grp = EC_KEY_get0_group(key);
-    if (!grp) return 0;
-
-    pt = EC_KEY_get0_public_key(key);
-    if (!pt) return 0;
-
-    meth = EC_GROUP_method_of(grp);
-    if (EC_METHOD_get_field_type(meth) == NID_X9_62_prime_field) {
-        rv = EC_POINT_get_affine_coordinates_GFp(grp, pt, x, y, ctx);
-    } else {
-        rv = EC_POINT_get_affine_coordinates_GF2m(grp, pt, x, y, ctx);
-    }
-
-    BN_CTX_free(ctx);
-    return rv;
-}
-
-/*
- * group values are declared near the top so that they
- * can be freed at the end of main execution
- */
-static int app_ecdsa_handler(ACVP_TEST_CASE *test_case) {
-    ACVP_ECDSA_TC    *tc;
-    int rv = 1;
-    ACVP_CIPHER mode;
-    const EVP_MD *md = NULL;
-    ECDSA_SIG *sig = NULL;
-
-    int nid = NID_undef, rc = 0, msg_len = 0;
-    BIGNUM *Qx = NULL, *Qy = NULL;
-    BIGNUM *r = NULL, *s = NULL;
-    const BIGNUM *d = NULL;
-    EC_KEY *key = NULL;
-
-
-    if (!test_case) {
-        printf("No test case found\n");
-        goto err;
-    }
-    tc = test_case->tc.ecdsa;
-    if (!tc) {
-        printf("\nError: test case not found in ECDSA handler\n");
-        goto err;
-    }
-    mode = tc->cipher;
-
-    if (mode == ACVP_ECDSA_SIGGEN || mode == ACVP_ECDSA_SIGVER) {
-        switch (tc->hash_alg) {
-        case ACVP_SHA1:
-            md = EVP_sha1();
-            break;
-        case ACVP_SHA224:
-            md = EVP_sha224();
-            break;
-        case ACVP_SHA256:
-            md = EVP_sha256();
-            break;
-        case ACVP_SHA384:
-            md = EVP_sha384();
-            break;
-        case ACVP_SHA512:
-            md = EVP_sha512();
-            break;
-        default:
-            printf("Unsupported hash alg in ECDSA\n");
-            goto err;
-        }
-    }
-
-    switch (tc->curve) {
-    case ACVP_EC_CURVE_B233:
-        nid = NID_sect233r1;
-        break;
-    case ACVP_EC_CURVE_B283:
-        nid = NID_sect283r1;
-        break;
-    case ACVP_EC_CURVE_B409:
-        nid = NID_sect409r1;
-        break;
-    case ACVP_EC_CURVE_B571:
-        nid = NID_sect571r1;
-        break;
-    case ACVP_EC_CURVE_K233:
-        nid = NID_sect233k1;
-        break;
-    case ACVP_EC_CURVE_K283:
-        nid = NID_sect283k1;
-        break;
-    case ACVP_EC_CURVE_K409:
-        nid = NID_sect409k1;
-        break;
-    case ACVP_EC_CURVE_K571:
-        nid = NID_sect571k1;
-        break;
-    case ACVP_EC_CURVE_P224:
-        nid = NID_secp224r1;
-        break;
-    case ACVP_EC_CURVE_P256:
-        nid = NID_X9_62_prime256v1;
-        break;
-    case ACVP_EC_CURVE_P384:
-        nid = NID_secp384r1;
-        break;
-    case ACVP_EC_CURVE_P521:
-        nid = NID_secp521r1;
-        break;
-    default:
-        printf("Unsupported curve\n");
-        goto err;
-    }
-
-    switch (mode) {
-    case ACVP_ECDSA_KEYGEN:
-        Qx = FIPS_bn_new();
-        Qy = FIPS_bn_new();
-        if (!Qx || !Qy) {
-            printf("Error BIGNUM malloc\n");
-            goto err;
-        }
-
-        key = EC_KEY_new_by_curve_name(nid);
-        if (!key) {
-            printf("Failed to instantiate ECDSA key\n");
-            goto err;
-        }
-
-        if (!EC_KEY_generate_key(key)) {
-            printf("Error generating ECDSA key\n");
-            goto err;
-        }
-
-        if (!ec_get_pubkey(key, Qx, Qy)) {
-            printf("Error getting ECDSA key attributes\n");
-            goto err;
-        }
-
-        d = EC_KEY_get0_private_key(key);
-
-        tc->qx_len = BN_bn2bin(Qx, tc->qx);
-        tc->qy_len = BN_bn2bin(Qy, tc->qy);
-        tc->d_len = BN_bn2bin(d, tc->d);
-        break;
-    case ACVP_ECDSA_KEYVER:
-        Qx = FIPS_bn_new();
-        Qy = FIPS_bn_new();
-        if (!tc->qx || !tc->qy) {
-            printf("missing qx or qy: ecdsa keyver\n");
-            goto err;
-        }
-        BN_bin2bn(tc->qx, tc->qx_len, Qx);
-        BN_bin2bn(tc->qy, tc->qy_len, Qy);
-        if (!Qx || !Qy) {
-            printf("Error BIGNUM conversion\n");
-            goto err;
-        }
-
-        key = EC_KEY_new_by_curve_name(nid);
-        if (!key) {
-            printf("Failed to instantiate ECDSA key\n");
-            goto err;
-        }
-
-        if (EC_KEY_set_public_key_affine_coordinates(key, Qx, Qy) == 1) {
-            tc->ver_disposition = ACVP_TEST_DISPOSITION_PASS;
-        } else {
-            tc->ver_disposition = ACVP_TEST_DISPOSITION_FAIL;
-        }
-        break;
-    case ACVP_ECDSA_SIGGEN:
-        if (ecdsa_current_tg != tc->tg_id) {
-            ecdsa_current_tg = tc->tg_id;
-            if (ecdsa_group_key) EC_KEY_free(ecdsa_group_key);
-
-            ecdsa_group_Qx = FIPS_bn_new();
-            ecdsa_group_Qy = FIPS_bn_new();
-            if (!ecdsa_group_Qx || !ecdsa_group_Qy) {
-                printf("Error BIGNUM malloc\n");
-                goto err;
-            }
-            ecdsa_group_key = EC_KEY_new_by_curve_name(nid);
-            if (!ecdsa_group_key) {
-                printf("Failed to instantiate ECDSA key\n");
-                goto err;
-            }
-
-            if (!EC_KEY_generate_key(ecdsa_group_key)) {
-                printf("Error generating ECDSA key\n");
-                goto err;
-            }
-
-            if (!ec_get_pubkey(ecdsa_group_key, ecdsa_group_Qx, ecdsa_group_Qy)) {
-                printf("Error getting ECDSA key attributes\n");
-                goto err;
-            }
-        }
-        msg_len = tc->msg_len;
-        if (!tc->message) {
-            printf("ecdsa siggen missing msg\n");
-            goto err;
-        }
-        sig = FIPS_ecdsa_sign(ecdsa_group_key, tc->message, msg_len, md);
-        if (!sig) {
-            printf("Error signing message\n");
-            goto err;
-        }
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-        r = sig->r;
-        s = sig->s;
-#else
-        ECDSA_SIG_get0(sig, (const BIGNUM **)&r,
-                       (const BIGNUM **)&s);
-#endif
-
-        tc->qx_len = BN_bn2bin(ecdsa_group_Qx, tc->qx);
-        tc->qy_len = BN_bn2bin(ecdsa_group_Qy, tc->qy);
-        tc->r_len = BN_bn2bin(r, tc->r);
-        tc->s_len = BN_bn2bin(s, tc->s);
-
-        break;
-    case ACVP_ECDSA_SIGVER:
-        if (!tc->message) {
-            printf("missing sigver message - nothing to verify\n");
-            goto err;
-        }
-        if (!tc->r) {
-            printf("missing r ecdsa sigver\n");
-            goto err;
-        }
-        if (!tc->s) {
-            printf("missing s ecdsa sigver\n");
-            goto err;
-        }
-        sig = ECDSA_SIG_new();
-        if (!sig) {
-            printf("Error generating ecdsa signature\n");
-            goto err;
-        }
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-        r = sig->r;
-        s = sig->s;
-#else
-        ECDSA_SIG_get0(sig, (const BIGNUM **)&r,
-                       (const BIGNUM **)&s);
-#endif
-
-        Qx = FIPS_bn_new();
-        Qy = FIPS_bn_new();
-
-        BN_bin2bn(tc->qx, tc->qx_len, Qx);
-        BN_bin2bn(tc->qy, tc->qy_len, Qy);
-        if (!Qx || !Qy) {
-            printf("Error BIGNUM conversion\n");
-            goto err;
-        }
-
-        BN_bin2bn(tc->r, tc->r_len, r);
-        BN_bin2bn(tc->s, tc->s_len, s);
-        if (!r || !s) {
-            printf("Error BIGNUM conversion\n");
-            goto err;
-        }
-
-        key = EC_KEY_new_by_curve_name(nid);
-        if (!key) {
-            printf("Failed to instantiate ECDSA key\n");
-            goto err;
-        }
-
-        rc = EC_KEY_set_public_key_affine_coordinates(key, Qx, Qy);
-        if (rc != 1) {
-            printf("Error setting ECDSA coordinates\n");
-            goto points_err;
-        }
-
-        if (!tc->message) {
-            printf("ecdsa siggen missing msg\n");
-            goto err;
-        }
-        if (FIPS_ecdsa_verify(key, tc->message, tc->msg_len, md, sig) == 1) {
-            tc->ver_disposition = ACVP_TEST_DISPOSITION_PASS;
-        } else {
-            tc->ver_disposition = ACVP_TEST_DISPOSITION_FAIL;
-        }
-points_err:
-        break;
-    default:
-        printf("Unsupported ECDSA mode\n");
-        break;
-    }
-    rv = 0;
-
-err:
-    if (sig) FIPS_ecdsa_sig_free(sig);
-    if (Qx) FIPS_bn_free(Qx);
-    if (Qy) FIPS_bn_free(Qy);
-    if (key) EC_KEY_free(key);
-    return rv;
-}
-
-/*
- * RSA SigGen handler
- * requires Makefile.fom to function
- *
- * group values are decalred near the top so that they
- * can be freed at the end of main execution
- */
-static int app_rsa_sig_handler(ACVP_TEST_CASE *test_case) {
-    EVP_MD *tc_md = NULL;
-    int siglen, pad_mode;
-    BIGNUM *bn_e = NULL, *e = NULL, *n = NULL;
-    ACVP_RSA_SIG_TC    *tc;
-    RSA *rsa = NULL;
-    int salt_len = -1;
-
-    int rv = 1;
-
-    if (!test_case) {
-        printf("\nError: test case not found in RSA SigGen handler\n");
-        goto err;
-    }
-
-    tc = test_case->tc.rsa_sig;
-
-    if (!tc) {
-        printf("\nError: test case not found in RSA SigGen handler\n");
-        goto err;
-    }
-
-    /*
-     * Make an RSA object and set a new BN exponent to use to generate a key
-     */
-
-    rsa = FIPS_rsa_new();
-    if (!rsa) {
-        printf("\nError: Issue with RSA obj in RSA Sig\n");
-        goto err;
-    }
-
-    bn_e = BN_new();
-    if (!bn_e || !BN_set_word(bn_e, 0x1001)) {
-        printf("\nError: Issue with exponent in RSA Sig\n");
-        goto err;
-    }
-
-    if (!tc->modulo) {
-        printf("\nError: Issue with modulo in RSA Sig\n");
-        goto err;
-    }
-
-    /*
-     * Set the pad mode and generate a key given the respective sigType
-     */
-    switch (tc->sig_type) {
-    case ACVP_RSA_SIG_TYPE_X931:
-        pad_mode = RSA_X931_PADDING;
-        salt_len = -2;
-        break;
-    case ACVP_RSA_SIG_TYPE_PKCS1PSS:
-        pad_mode = RSA_PKCS1_PSS_PADDING;
-        salt_len = tc->salt_len;
-        break;
-    case ACVP_RSA_SIG_TYPE_PKCS1V15:
-        pad_mode = RSA_PKCS1_PADDING;
-        break;
-    default:
-        printf("\nError: sigType not supported\n");
-        rv = ACVP_INVALID_ARG;
-        goto err;
-    }
-
-    /*
-     * Set the message digest to the appropriate sha
-     */
-    switch (tc->hash_alg) {
-    case ACVP_SHA1:
-        tc_md = (EVP_MD *)EVP_sha1();
-        break;
-    case ACVP_SHA224:
-        tc_md = (EVP_MD *)EVP_sha224();
-        break;
-    case ACVP_SHA256:
-        tc_md = (EVP_MD *)EVP_sha256();
-        break;
-    case ACVP_SHA384:
-        tc_md = (EVP_MD *)EVP_sha384();
-        break;
-    case ACVP_SHA512:
-        tc_md = (EVP_MD *)EVP_sha512();
-        break;
-    default:
-        printf("\nError: hashAlg not supported for RSA SigGen\n");
-        goto err;
-    }
-
-    /*
-     * If we are verifying, set RSA to the given public key
-     * Else, generate a new key, retrieve and save values
-     */
-    if (tc->sig_mode == ACVP_RSA_SIGVER) {
-        e = BN_new();
-        if (!e) {
-            printf("\nBN alloc failure (e)\n");
-            goto err;
-        }
-        BN_bin2bn(tc->e, tc->e_len, e);
-
-        n = BN_new();
-        if (!n) {
-            printf("\nBN alloc failure (n)\n");
-            goto err;
-        }
-        BN_bin2bn(tc->n, tc->n_len, n);
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-        rsa->e = BN_dup(e);
-        rsa->n = BN_dup(n);
-#else
-        RSA_set0_key(rsa, n, e, NULL);
-#endif
-
-        tc->ver_disposition = FIPS_rsa_verify(rsa, tc->msg, tc->msg_len, tc_md, pad_mode, salt_len, NULL, tc->signature, tc->sig_len);
-    } else {
-        if (rsa_current_tg != tc->tg_id) {
-            rsa_current_tg = tc->tg_id;
-            if (group_rsa) RSA_free(group_rsa);
-            group_rsa = RSA_new();
-            if (!FIPS_rsa_x931_generate_key_ex(group_rsa, tc->modulo, bn_e, NULL)) {
-                printf("\nError: Issue with keygen during siggen handling\n");
-                goto err;
-            }
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-            e = group_rsa->e;
-            n = group_rsa->n;
-            group_n = BN_dup(n);
-#else
-            RSA_get0_key(rsa, (const BIGNUM **)&n, (const BIGNUM **)&e, NULL);
-#endif
-        } else {
-            e = BN_dup(bn_e);
-            n = BN_dup(group_n);
-        }
-        tc->e_len = BN_bn2bin(e, tc->e);
-        tc->n_len = BN_bn2bin(n, tc->n);
-
-        if (tc->msg && tc_md) {
-            siglen = RSA_size(group_rsa);
-
-            if (!FIPS_rsa_sign(group_rsa, tc->msg, tc->msg_len, tc_md, pad_mode, salt_len, NULL,
-                               tc->signature, (unsigned int *)&siglen)) {
-                printf("\nError: RSA Signature Generation fail\n");
-                goto err;
-            }
-
-            tc->sig_len = siglen;
-        }
-    }
-    rv = 0;
-err:
-    if (bn_e) BN_free(bn_e);
-    if (rsa) FIPS_rsa_free(rsa);
-
-    return rv;
-}
-#endif
-
-#ifdef ACVP_NO_RUNTIME
-typedef struct {
-    unsigned char *ent;
-    size_t entlen;
-    unsigned char *nonce;
-    size_t noncelen;
-} DRBG_TEST_ENT;
-
-static size_t drbg_test_entropy(DRBG_CTX *dctx,
-                                unsigned char **pout,
-                                int entropy,
-                                size_t min_len,
-                                size_t max_len) {
-    if (!dctx || !pout) return 0;
-
-    DRBG_TEST_ENT *t = (DRBG_TEST_ENT *)FIPS_drbg_get_app_data(dctx);
-    if (!t) return 0;
-
-    if (t->entlen < min_len) printf("entropy data len < min_len: %zu\n", t->entlen);
-    if (t->entlen > max_len) printf("entropy data len > max_len: %zu\n", t->entlen);
-    *pout = (unsigned char *)t->ent;
-    return t->entlen;
-}
-
-static size_t drbg_test_nonce(DRBG_CTX *dctx,
-                              unsigned char **pout,
-                              int entropy,
-                              size_t min_len,
-                              size_t max_len) {
-    if (!dctx || !pout) return 0;
-
-    DRBG_TEST_ENT *t = (DRBG_TEST_ENT *)FIPS_drbg_get_app_data(dctx);
-
-    if (t->noncelen < min_len) printf("nonce data len < min_len: %zu\n", t->noncelen);
-    if (t->noncelen > max_len) printf("nonce data len > max_len: %zu\n", t->noncelen);
-    *pout = (unsigned char *)t->nonce;
-    return t->noncelen;
-}
-
-static int app_drbg_handler(ACVP_TEST_CASE *test_case) {
-    int result = 1;
-    ACVP_DRBG_TC    *tc;
-    unsigned int nid;
-    int der_func = 0;
-    unsigned int drbg_entropy_len;
-    int fips_rc;
-
-    unsigned char   *nonce = NULL;
-
-    if (!test_case) {
-        return result;
-    }
-
-    tc = test_case->tc.drbg;
-    /*
-     * Init entropy length
-     */
-    drbg_entropy_len = tc->entropy_len;
-
-    switch (tc->cipher) {
-    case ACVP_HASHDRBG:
-        nonce = tc->nonce;
-        switch (tc->mode) {
-        case ACVP_DRBG_SHA_1:
-            nid = NID_sha1;
-            break;
-        case ACVP_DRBG_SHA_224:
-            nid = NID_sha256;
-            break;
-        case ACVP_DRBG_SHA_256:
-            nid = NID_sha256;
-            break;
-        case ACVP_DRBG_SHA_384:
-            nid = NID_sha384;
-            break;
-        case ACVP_DRBG_SHA_512:
-            nid = NID_sha512;
-            break;
-
-        case ACVP_DRBG_SHA_512_224:
-        case ACVP_DRBG_SHA_512_256:
-        default:
-            printf("%s: Unsupported algorithm/mode %d/%d (tc_id=%d)\n", __FUNCTION__, tc->tc_id,
-                   tc->cipher, tc->mode);
-            return result;
-
-            break;
-        }
-        break;
-
-    case ACVP_HMACDRBG:
-        nonce = tc->nonce;
-        switch (tc->mode) {
-        case ACVP_DRBG_SHA_1:
-            nid =   NID_hmacWithSHA1;
-            break;
-        case ACVP_DRBG_SHA_224:
-            nid =   NID_hmacWithSHA224;
-            break;
-        case ACVP_DRBG_SHA_256:
-            nid =   NID_hmacWithSHA256;
-            break;
-        case ACVP_DRBG_SHA_384:
-            nid =   NID_hmacWithSHA384;
-            break;
-        case ACVP_DRBG_SHA_512:
-            nid =   NID_hmacWithSHA512;
-            break;
-        case ACVP_DRBG_SHA_512_224:
-        case ACVP_DRBG_SHA_512_256:
-        default:
-            printf("%s: Unsupported algorithm/mode %d/%d (tc_id=%d)\n", __FUNCTION__, tc->tc_id,
-                   tc->cipher, tc->mode);
-            return result;
-
-            break;
-        }
-        break;
-
-    case ACVP_CTRDRBG:
-        /*
-         * DR function Only valid in CTR mode
-         * if not set nonce is ignored
-         */
-        if (tc->der_func_enabled) {
-            der_func = DRBG_FLAG_CTR_USE_DF;
-            nonce = tc->nonce;
-        } else {
-            /**
-             * Note 5: All DRBGs are tested at their maximum supported security
-             * strength so this is the minimum bit length of the entropy input that
-             * ACVP will accept.  The maximum supported security strength is also
-             * the default value for this input.  Longer entropy inputs are
-             * permitted, with the following exception: for ctrDRBG with no df, the
-             * bit length must equal the seed length.
-             **/
-            drbg_entropy_len = 0;
-        }
-
-        switch (tc->mode) {
-        case ACVP_DRBG_AES_128:
-            nid = NID_aes_128_ctr;
-            break;
-        case ACVP_DRBG_AES_192:
-            nid = NID_aes_192_ctr;
-            break;
-        case ACVP_DRBG_AES_256:
-            nid = NID_aes_256_ctr;
-            break;
-        case ACVP_DRBG_3KEYTDEA:
-        default:
-            printf("%s: Unsupported algorithm/mode %d/%d (tc_id=%d)\n", __FUNCTION__, tc->tc_id,
-                   tc->cipher, tc->mode);
-            return result;
-
-            break;
-        }
-        break;
-    default:
-        printf("%s: Unsupported algorithm %d (tc_id=%d)\n", __FUNCTION__, tc->tc_id,
-               tc->cipher);
-        return result;
-
-        break;
-    }
-
-    DRBG_CTX *drbg_ctx = NULL;
-    DRBG_TEST_ENT entropy_nonce;
-    memzero_s(&entropy_nonce, sizeof(DRBG_TEST_ENT));
-    drbg_ctx = FIPS_drbg_new(nid, der_func | DRBG_FLAG_TEST);
-    if (!drbg_ctx) {
-        progress("ERROR: failed to create DRBG Context.");
-        return result;
-    }
-
-    /*
-     * Set entropy and nonce
-     */
-    entropy_nonce.ent = tc->entropy;
-    entropy_nonce.entlen = drbg_entropy_len;
-
-    entropy_nonce.nonce = nonce;
-    entropy_nonce.noncelen = tc->nonce_len;
-
-    FIPS_drbg_set_app_data(drbg_ctx, &entropy_nonce);
-
-    fips_rc = FIPS_drbg_set_callbacks(drbg_ctx,
-                                      drbg_test_entropy,
-                                      0, 0,
-                                      drbg_test_nonce,
-                                      0);
-    if (!fips_rc) {
-        progress("ERROR: failed to Set callback DRBG ctx");
-        long l = 9;
-        char buf[2048]  = { 0 };
-        while ((l = ERR_get_error())) {
-            printf("ERROR:%s\n", ERR_error_string(l, buf));
-        }
-        goto end;
-    }
-
-    fips_rc = FIPS_drbg_instantiate(drbg_ctx, (const unsigned char *)tc->perso_string,
-                                    (size_t)tc->perso_string_len);
-    if (!fips_rc) {
-        progress("ERROR: failed to instantiate DRBG ctx");
-        long l = 9;
-        char buf[2048]  = { 0 };
-        while ((l = ERR_get_error())) {
-            printf("ERROR:%s\n", ERR_error_string(l, buf));
-        }
-        goto end;
-    }
-
-    /*
-     * Process predictive resistance flag
-     */
-    if (tc->pred_resist_enabled) {
-        entropy_nonce.ent = tc->entropy_input_pr;
-        entropy_nonce.entlen = drbg_entropy_len;
-
-        fips_rc =  FIPS_drbg_generate(drbg_ctx, (unsigned char *)tc->drb,
-                                      (size_t)(tc->drb_len),
-                                      (int)1,
-                                      (const unsigned char *)tc->additional_input,
-                                      (size_t)(tc->additional_input_len));
-        if (!fips_rc) {
-            progress("ERROR: failed to generate drb");
-            long l;
-            while ((l = ERR_get_error())) {
-                printf("ERROR:%s\n", ERR_error_string(l, NULL));
-            }
-            goto end;
-        }
-
-        entropy_nonce.ent = tc->entropy_input_pr_1;
-        entropy_nonce.entlen = drbg_entropy_len;
-
-        fips_rc =  FIPS_drbg_generate(drbg_ctx, (unsigned char *)tc->drb,
-                                      (size_t)(tc->drb_len),
-                                      (int)1,
-                                      (const unsigned char *)tc->additional_input_1,
-                                      (size_t)(tc->additional_input_len));
-        if (!fips_rc) {
-            progress("ERROR: failed to generate drb");
-            long l;
-            while ((l = ERR_get_error())) {
-                printf("ERROR:%s\n", ERR_error_string(l, NULL));
-            }
-            goto end;
-        }
-    } else {
-        fips_rc = FIPS_drbg_generate(drbg_ctx, (unsigned char *)tc->drb,
-                                     (size_t)(tc->drb_len),
-                                     (int)0,
-                                     (const unsigned char *)tc->additional_input,
-                                     (size_t)(tc->additional_input_len));
-        if (!fips_rc) {
-            progress("ERROR: failed to generate drb");
-            long l;
-            while ((l = ERR_get_error())) {
-                printf("ERROR:%s\n", ERR_error_string(l, NULL));
-            }
-            goto end;
-        }
-    }
-    result = 0;
-
-end:
-    FIPS_drbg_uninstantiate(drbg_ctx);
-    FIPS_drbg_free(drbg_ctx);
-
-    return result;
-}
-#endif
-
-
-/* This is a public domain base64 implementation written by WEI Zhicheng. */
-
-enum { BASE64_OK = 0, BASE64_INVALID };
-
-#define BASE64_ENCODE_OUT_SIZE(s)    (((s) + 2) / 3 * 4)
-#define BASE64_DECODE_OUT_SIZE(s)    (((s)) / 4 * 3)
-
-#define BASE64_PAD    '='
-
-
-#define BASE64DE_FIRST    '+'
-#define BASE64DE_LAST    'z'
-/* ASCII order for BASE 64 decode, -1 in unused character */
-static const signed char base64de[] = {
-    /* '+', ',', '-', '.', '/', '0', '1', '2', */
-    62, -1, -1, -1, 63, 52, 53, 54,
-
-    /* '3', '4', '5', '6', '7', '8', '9', ':', */
-    55, 56, 57, 58, 59, 60, 61, -1,
-
-    /* ';', '<', '=', '>', '?', '@', 'A', 'B', */
-    -1, -1, -1, -1, -1, -1, 0,  1,
-
-    /* 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', */
-    2,  3,  4,  5,  6,  7,  8,  9,
-
-    /* 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', */
-    10, 11, 12, 13, 14, 15, 16, 17,
-
-    /* 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z', */
-    18, 19, 20, 21, 22, 23, 24, 25,
-
-    /* '[', '\', ']', '^', '_', '`', 'a', 'b', */
-    -1, -1, -1, -1, -1, -1, 26, 27,
-
-    /* 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', */
-    28, 29, 30, 31, 32, 33, 34, 35,
-
-    /* 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', */
-    36, 37, 38, 39, 40, 41, 42, 43,
-
-    /* 's', 't', 'u', 'v', 'w', 'x', 'y', 'z', */
-    44, 45, 46, 47, 48, 49, 50, 51,
-};
-
-static unsigned int
-base64_decode(const char *in, unsigned int inlen, unsigned char *out) {
-    unsigned int i, j;
-
-    for (i = j = 0; i < inlen; i++) {
-        int c;
-        int s = i % 4;             /* from 8/gcd(6, 8) */
-
-        if (in[i] == '=')
-            return j;
-
-        if (in[i] < BASE64DE_FIRST || in[i] > BASE64DE_LAST ||
-            (c = base64de[in[i] - BASE64DE_FIRST]) == -1)
-            return 0;
-
-        switch (s) {
-        case 0:
-            out[j] = ((unsigned int)c << 2) & 0xFF;
-            continue;
-        case 1:
-            out[j++] += ((unsigned int)c >> 4) & 0x3;
-
-            /* if not last char with padding */
-            if (i < (inlen - 3) || in[inlen - 2] != '=')
-                out[j] = ((unsigned int)c & 0xF) << 4;
-            continue;
-        case 2:
-            out[j++] += ((unsigned int)c >> 2) & 0xF;
-
-            /* if not last char with padding */
-            if (i < (inlen - 2) || in[inlen - 1] != '=')
-                out[j] =  ((unsigned int)c & 0x3) << 6;
-            continue;
-        case 3:
-            out[j++] += (unsigned char)c;
-        }
-    }
-
-    return j;
-}
-
-#define T_LEN 8
-#define MAX_LEN 512
-
-const int DIGITS_POWER[]
-//  0  1   2    3     4      5       6        7         8
-    = { 1, 10, 100, 1000, 10000, 100000, 1000000, 10000000, 100000000 };
-
-static
-int hmac_totp(const char *key,
-              const unsigned char *msg,
-              char *hash,
-              int hash_max,
-              const EVP_MD *md,
-              unsigned int key_len) {
-    int len = 0;
-    unsigned char buff[MAX_LEN];
-    HMAC_CTX *ctx;
-
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-    HMAC_CTX static_ctx;
-
-    ctx = &static_ctx;
-    HMAC_CTX_init(ctx);
-#else
-    ctx = HMAC_CTX_new();
-#endif
-
-    HMAC_CTX_set_flags(ctx, EVP_MD_CTX_FLAG_NON_FIPS_ALLOW);
-    if (!HMAC_Init_ex(ctx, key, key_len, md, NULL)) goto end;
-    if (!HMAC_Update(ctx, msg, T_LEN)) goto end;
-    if (!HMAC_Final(ctx, buff, (unsigned int *)&len)) goto end;
-    memcpy_s(hash, hash_max, buff, len);
-
-end:
-#if OPENSSL_VERSION_NUMBER <= 0x10100000L
-    HMAC_CTX_cleanup(ctx);
-#else
-    if (ctx) HMAC_CTX_free(ctx);
-#endif
-
-    return len;
-}
-
-static ACVP_RESULT totp(char **token, int token_max) {
-    char hash[MAX_LEN] = {0};
-    int os, bin, otp;
-    int md_len;
-    char format[5];
-    time_t t;
-    unsigned char token_buff[T_LEN + 1] = {0};
-    char *new_seed = calloc(ACVP_TOTP_TOKEN_MAX, sizeof(char));
-    char *seed = NULL;
-    int seed_len = 0;
-
-    if (!new_seed) {
-        printf("Failed to malloc new_seed\n");
-        return ACVP_MALLOC_FAIL;
-    }
-
-    seed = getenv("ACV_TOTP_SEED");
-    if (!seed) {
-        printf("Failed to get TOTP seed\n");
-        free(new_seed);
-        return ACVP_TOTP_MISSING_SEED;
-    }
-
-    t = time(NULL);
-
-    // RFC4226
-    t = t / 30;
-    token_buff[0] = (t >> T_LEN * 7) & 0xff;
-    token_buff[1] = (t >> T_LEN * 6) & 0xff;
-    token_buff[2] = (t >> T_LEN * 5) & 0xff;
-    token_buff[3] = (t >> T_LEN * 4) & 0xff;
-    token_buff[4] = (t >> T_LEN * 3) & 0xff;
-    token_buff[5] = (t >> T_LEN * 2) & 0xff;
-    token_buff[6] = (t >> T_LEN * 1) & 0xff;
-    token_buff[7] = t & 0xff;
-
-#define MAX_SEED_LEN 64
-    seed_len = base64_decode(seed, strnlen_s(seed, MAX_SEED_LEN), (unsigned char *)new_seed);
-    if (seed_len  == 0) {
-        printf("Failed to decode TOTP seed\n");
-        free(new_seed);
-        return ACVP_TOTP_DECODE_FAIL;
-    }
-
-
-    // use passed hash function
-    md_len = hmac_totp(new_seed, token_buff, hash, sizeof(hash), EVP_sha256(), seed_len);
-    if (md_len == 0) {
-        printf("Failed to create TOTP\n");
-        free(new_seed);
-        return ACVP_CRYPTO_MODULE_FAIL;
-    }
-    os = hash[(int)md_len - 1] & 0xf;
-
-    bin = ((hash[os + 0] & 0x7f) << 24) |
-          ((hash[os + 1] & 0xff) << 16) |
-          ((hash[os + 2] & 0xff) <<  8) |
-          ((hash[os + 3] & 0xff) <<  0);
-
-    otp = bin % DIGITS_POWER[ACVP_TOTP_LENGTH];
-
-    // generate format string like "%06d" to fix digits using 0
-    sprintf(format, "%c0%ldd", '%', (long int)ACVP_TOTP_LENGTH);
-
-    sprintf((char *)token_buff, format, otp);
-    memcpy_s((char *)*token, token_max, token_buff, ACVP_TOTP_LENGTH);
-    free(new_seed);
-    return ACVP_SUCCESS;
-}
